==> notebooks/1-1-Genomeannotation.ipynb <==
# 1-1-Genomeannotation
Jakke Neiro$^1$
1. Aboobaker lab, Department of Zoology, University of Oxford

## Contents of notebook
* 1. Introduction
* 2. Data quality
* 3. Alignment
* 4. Transcript assembly and merge

## Files
* Input: 183 RNA-seq samples
* Output: stringtie_merged.gtf

# 1. Introduction

This notebook describes the annotation of the sexual planarian genome (*Schmidtea mediterranea*) based on total RNA-seq data from this species. In short, took advantage of the Stringtie annotation pipeline by using 183 planarian RNA-seq reads as the raw RNA-seq material and the PlanMine sexual genome annotation (SMESG-high) as the reference annotation. More information on the stringtie annotation process can be found at http://ccb.jhu.edu/software/stringtie/index.shtml.

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro
grep -v -P '\t\.\t\.\t' stringtie_merged.gtf > stringtie.Bioconductor.gtf

# 2. Data quality

In total 183 RNA-seq data sets sequenced in the planarian were used as the raw pool of RNA-seq reads for the expression-driven annotation. The sample IDs are listed below:

%%bash
cd /drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/sexual_genome_annotation/
wc -l sample_ids.txt
less sample_ids.txt

# 3. Alignment

The SMESG-high genome annotation and gene models were downloaded from PlanMine v3.0.

#%%bash
#cd ../sexual_genome_annotation_files/sexual_genome_annotation/
#wget http://planmine.mpi-cbg.de/planmine/model/bulkdata/smes_v2_hconf_SMESG.gff3.zip
#unzip smes_v2_hconf_SMESG.gff3.zip

The gff3 file was converted into a gtf file:

%%bash
cd ../sexual_genome_annotation_files/sexual_genome_annotation/
gffread smes_v2_hconf_SMESG.gff3 -T -o smes_v2_hconf_SMESG.gtf

The splice sites were extracted for hisat2:

%%bash
cd ../sexual_genome_annotation_files/sexual_genome_annotation/
~/software/hisat2-2.1.0/hisat2_extract_splice_sites.py smes_v2_hconf_SMESG.gtf > smes_v2_hconf_hisat2_splice_sites

The exon sites were extracted for hisat2:

%%bash
cd ../sexual_genome_annotation_files/sexual_genome_annotation/
~/software/hisat2-2.1.0/hisat2_extract_exons.py smes_v2_hconf_SMESG.gtf > smes_v2_hconf_hisat2_exons

The hisat2 index was build for alignment:

%%bash
cd ../sexual_genome_annotation_files/sexual_genome_annotation/
nohup ~/software/hisat2-2.1.0/hisat2-build final_dd_Smed_g4.fa final_dd_Smed_g4_hisat2 > nohup_hisat_index &

All RNA-seq libraries were aligned to the sexual planarian genome with HISAT2:

%%bash
cd /hydra/neiro_bam
less hisat2_align.sh

#%%bash
#cd /hydra/neiro_bam
#nohup ./hisat2_align.sh &

Finally, all the sam files were converted into bam files:

%%bash
cd /hydra/neiro_bam
less hisat_samtools.sh

#%%bash
#nohup ./hisat_samtools.sh &

# 4. Transcript assembly and merge
The assembly process used Stringtie. First, the RNA-seq reads were assembled into transcripts for each sample using the exisiting SMESG-high annotation.

%%bash
cd /hydra/neiro_bam
less stringtie.sh

#%%bash
#cd /hydra/neiro_bam
#nohup ./stringtie.sh &

All the bam files generated in the previous step were collected into a single list:

%%bash
cd /hydra/neiro_bam
for f in *.gtf; do echo $f >> mergelist.txt; done

Finally, all the sample-specific annotations were merged into one final annotation.

%%bash
cd /hydra/neiro_bam
less stringtie_merge.sh

%%bash
cd /hydra/neiro_bam
nohup ./stringtie_merge.sh &

# FINNISHED
