==> notebooks/1-2-Genome-annotation-analysis.ipynb <==
# 1-2-Genome-annotation-analysis
Jakke Neiro$^1$
1. Aboobaker laboratory, Department of Zoology, University of Oxford

## Contents of notebook
* 1. Introduction
* 2. Comparison of annotations
    * 2.1 Rink High-confidence vs. Neiro
* 3. GO terms for Rink high-confidence transcripts
    * 3.1 High-confidence transcripts from annotation file
    * 3.2 Transdecoder
    * 3.3 Interproscan
    * 3.4 GO terms
* 4. Coding and non-coding transcripts from new annotation
    * 4.1 Putative ncRNAs from annotation file
    * 4.2 Transdecoder
        * 4.2.1 Open reading frames
        * 4.2.2 Preparation for Interproscan
    * 4.3 Inteproscan
        * 4.3.1 Sequences longer than 100 amino acids
        * 4.3.2 Sequences shorter than 1000 amino acids
    * 4.4 Characterisation of transcripts
        * 4.4.1 Coding sequences longer than 100 amino acids
        * 4.4.2 Coding sequences shorter than 100 amino acids
        * 4.4.3 Non-coding transcripts

## Files:
* Input: stringtie_merged.gtf
* Output: interpro_rink.GO.tab, ncrna_coding_May2021.GO.tab, ncrna_new_coding_May2021.GO.tab, noncoding_final_May2021.csv

# 1. Introduction

The genome annotation generated in 1-1-Genomeannotation was used to identify new coding and non-coding transcripts. In addition, GO terms were assigned to transcripts from the Rink high-confidence transcripts. 

# 2. Comparison of annotations

First, the new annotation was compared to previous annotations. 

## 2.1 Rink High-confidence vs. Neiro

The new annotation ("Neiro") was compared to the Rink's high-confidence annotation from Planmine ("Rink High-confidence")(http://planmine.mpi-cbg.de/planmine/aspect.do?name=Gene%20Predictions) with gffcompare. 

#%%bash
#cd /hydra/neiro_bam
#/hydra/software/gffcompare-0.11.5/gffcompare -r ~/sexual_genome_annotation_files/sexual_genome_annotation/smes_v2_hconf_SMESG.gtf -G stringtie_merged.gtf

The gffcompare results were investigated:

%%bash
cd /hydra/neiro_bam
less gffcmp.stats

%%bash
cd /hydra/neiro_bam
head -2 gffcmp.tracking

%%bash
cd /hydra/neiro_bam
head -2 gffcmp.loci

%%bash
cd /hydra/neiro_bam
head -2 gffcmp.stringtie_merged.gtf.refmap

%%bash
cd /hydra/neiro_bam
head -2 gffcmp.stringtie_merged.gtf.tmap

%%bash
cd /hydra/neiro_bam
head -2 gffcmp.annotated.gtf

%%bash
cd /hydra/neiro_bam
grep -v -P '\t\.\t\.\t' stringtie_merged.gtf > stringtie_merged_R.gtf

28097-22129

# 3. GO terms for Rink high-confidence transcripts

The GO terms for Rink's high-confidence transcripts were extracted with Interproscan. The GO terms of these transcripts were later compared to the new identified transcripts (1-3-Genome-annotation-analysis-R, 3. Gene ontology analysis of new coding transcripts).   

## 3.1 High-confidence transcripts from annotation file

The high-confidence transcripts were extracted from the annotation file.

%%bash
cd /hydra/GO
gffread -x final_dd_Smed_g4_transcripts.fa -g ~/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa ~/sexual_genome_annotation_files/sexual_genome_annotation/smes_v2_hconf_SMESG.gff3

%%bash
cd /hydra/GO
gffread --version

%%bash
cd /hydra/GO
grep -c ">" final_dd_Smed_g4_transcripts.fa

## 3.2 Transdecoder

The open reading frames were computed with Transdecoder.

#%%bash
#cd /hydra/GO
#TransDecoder.LongOrfs -t final_dd_Smed_g4_transcripts.fa -o Transdecoder_final

%%bash
cd /hydra/GO
TransDecoder.LongOrfs --version

%%bash
cd /hydra/GO/Transdecoder_final
grep -c ">" longest_orfs.pep

The last "\*" character of the peptide sequences was removed.

%%bash
cd /hydra/GO/Transdecoder_final
cp longest_orfs.pep longest_orfs_ip.pep
sed -i 's_*__g' longest_orfs_ip.pep
head longest_orfs_ip.pep

## 3.3 Interproscan

The peptide file generated with Transdecoder was split up into parts for Interproscan:

%%bash
cd /hydra/GO/Transdecoder_final
seqkit range -r 1:5000 longest_orfs_ip.pep > orfs_1_5k.fa.pep
seqkit range -r 5001:10000 longest_orfs_ip.pep > orfs_5k_10k.fa.pep
seqkit range -r 10001:15000 longest_orfs_ip.pep > orfs_10k_15k.fa.pep
seqkit range -r 15001:20000 longest_orfs_ip.pep > orfs_15k_20k.fa.pep
seqkit range -r 20001:25000 longest_orfs_ip.pep > orfs_20k_25k.fa.pep
seqkit range -r 25001:30000 longest_orfs_ip.pep > orfs_25k_30k.fa.pep
seqkit range -r 30001:36368 longest_orfs_ip.pep > orfs_30k_36k.fa.pep

The GO terms (and other peptide domain information) were obtained with Interproscan: 

#%%bash
#cd /hydra/GO
#nohup ~/software/interproscan-5.39-77.0/interproscan.sh -i ./Transdecoder_final/orfs_1_5k.fa.pep -d interpro -appl Pfam, SUPERFAMILY -goterms &
#nohup ~/software/interproscan-5.39-77.0/interproscan.sh -i ./Transdecoder_final/orfs_5k_10k.fa.pep -d interpro -appl Pfam, SUPERFAMILY -goterms &
#nohup ~/software/interproscan-5.39-77.0/interproscan.sh -i ./Transdecoder_final/orfs_10k_15k.fa.pep -d interpro -appl Pfam, SUPERFAMILY -goterms &
#nohup ~/software/interproscan-5.39-77.0/interproscan.sh -i ./Transdecoder_final/orfs_15k_20k.fa.pep -d interpro -appl Pfam, SUPERFAMILY -goterms &
#nohup ~/software/interproscan-5.39-77.0/interproscan.sh -i ./Transdecoder_final/orfs_20k_25k.fa.pep -d interpro -appl Pfam, SUPERFAMILY -goterms &
#nohup ~/software/interproscan-5.39-77.0/interproscan.sh -i ./Transdecoder_final/orfs_25k_30k.fa.pep -d interpro -appl Pfam, SUPERFAMILY -goterms &
#nohup ~/software/interproscan-5.39-77.0/interproscan.sh -i ./Transdecoder_final/orfs_30k_36k.fa.pep -d interpro -appl Pfam, SUPERFAMILY -goterms &

## 3.4 GO terms

The GO terms were extracted from the Interproscan output file and retabulated (interpro_rink.GO.tab). 

%%bash
cd /hydra/GO/interpro
ls *tsv

paths = ["../GO/interpro/orfs_1_5k.fa.pep.tsv", "../GO/interpro/orfs_5k_10k.fa.pep.tsv", "../GO/interpro/orfs_10k_15k.fa.pep.tsv", "../GO/interpro/orfs_15k_20k.fa.pep.tsv", "../GO/interpro/orfs_20k_25k.fa.pep.tsv", "../GO/interpro/orfs_25k_30k.fa.pep.tsv"]
transcripts = []
gos = []
for path in paths:
    f = open(path)
    content = f.readlines()
    print(len(content))
    for i in range(0, len(content)):
        if len(content[i].split("\t")) == 14 and content[i].split("\t")[13] != "\n":
            transcript = content[i].split("\t")[0].rpartition(".")[0]
            go = content[i].split("\t")[13].split("\n")[0].split("|")
            #transcripts were only included once
            if transcript not in transcripts:
                transcripts.append(transcript)
                gos.append(go)

f = open("../GO/interpro/interpro_rink.GO.tab", "a+")
for i in range(0, len(transcripts)):
    f.write(str(transcripts[i])+"\t")
    for j in range(0, len(gos[i])):
        if j == (len(gos[i])-1):
            f.write(str(gos[i][j])+"\n")
        else:
            f.write(str(gos[i][j])+", ")

%%bash
cd /hydra/GO/interpro/
head -2 interpro_rink.GO.tab
wc -l interpro_rink.GO.tab

# 4. Coding and non-coding transcripts from new annotation

The transcripts from the new annotation were characterised.

## 4.1 Putative ncRNAs from annotation file

The gffcompare tmap file was imported:

import pandas as pd
gffcmp = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/gffcmp.stringtie_merged.gtf.tmap", sep="\t")

The total number of transcripts in the new annotation was calculated:

len(gffcmp)

The total number

len(gffcmp[gffcmp["class_code"] == "="])

ncrna = gffcmp[gffcmp["class_code"].isin(['u', 'i', "x"])]

len(ncrna)

len(gffcmp[gffcmp["class_code"].isin(['c', 'k', 'm', 'n', 'j', 'e', 'o'])])

The IDs of the putative ncRNAs were saves as a list file:

#ncrna.iloc[:, 4].to_csv("../sexual_genome_annotation_files/ncrna_Neiro/ncrna_list.csv", sep ="\t", index=False)

The ncRNAs were extracted as a gtf file:

#%%bash
#cd ../../../../../../data/ngs/smed/facs_rnaseq_datasets/sexual/neiro_bam
#grep -F -f ~/sexual_genome_annotation_files/ncrna_Neiro/ncrna_list.csv gffcmp.annotated.gtf > ~/sexual_genome_annotation_files/ncrna_Neiro/stringtie.annotated.ncrna.gtf

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro
head -2 stringtie.annotated.ncrna.gtf

The ncRNA sequences were extracted using the gtf file:

#%%bash
#cd ~/sexual_genome_annotation_files/ncrna_Neiro
#gffread -w ncrna_transcripts.fa -g ../sexual_genome_annotation/final_dd_Smed_g4.fa stringtie.annotated.ncrna.gtf

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro
head -2 ncrna_transcripts.fa

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro
grep -c ">" ncrna_transcripts.fa

Because of the grep function, some false postivies were included in the fasta file.

## 4.2 Transdecoder

### 4.2.1 Open reading frames

The open reading frames for new transcripts were computed with Transdecoder:

#%%bash
#cd ~/sexual_genome_annotation_files/ncrna_Neiro
#TransDecoder.LongOrfs -t ncrna_transcripts.fa -m 10 -O Transdecoder_final

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
grep -c ">" longest_orfs.pep

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
head -2 longest_orfs.gff3

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
head -2 longest_orfs.pep

Transcripts without an open reading frame were extracted:

import pandas as pd
orf10 = pd.read_csv("../sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/longest_orfs.gff3", sep="\t", header=None)

non_orf = []
orfs = pd.Categorical(orf10.iloc[:,0]).categories
for i in range(len(ncrna_list)):
    if ncrna_list.iloc[i, 0] not in orfs:
          non_orf.append(ncrna_list.iloc[i, 0])

pd.Series(non_orf).to_csv("../sexual_genome_annotation_files/ncrna_Neiro/noncoding10aa_list.csv"

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/
wc -l noncoding10aa_list.csv

### 4.2.2 Preparation for Interproscan

The peptide fasta file was split up into two parts: sequences longer than 100 amino acids and sequences shorter than 100 amino acids. 

%%bash
cd ~/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
seqkit seq -m 100 longest_orfs.pep > aa100longer_merged.fa.pep

%%bash
cd ~/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
seqkit seq -M 100 longest_orfs.pep > aa100shorter_merged.fa.pep

#%%bash
#cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
#head -2 aa100longer_merged.fa.pep

#%%bash
#cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
#head aa100shorter_merged.fa.pep

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
grep -c ">" aa100longer_merged.fa.pep

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
grep -c ">" aa100shorter_merged.fa.pep

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
#cp aa100longer_merged.fa.pep aa100longer_merged_ip.fa.pep
#sed -i 's_*__g' aa100longer_merged_ip.fa.pep
head -2 aa100longer_merged_ip.fa.pep

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
#cp aa100shorter_merged.fa.pep aa100shorter_merged_ip.fa.pep
#sed -i 's_*__g' aa100shorter_merged_ip.fa.pep
head -2 aa100shorter_merged_ip.fa.pep

## 4.3 Interproscan

### 4.3.1 Sequences longer than 100 amino acids

Interproscan was run on sequences longer than 100 amino acids.

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/
echo "#!/bin/bash" > interproscan_more100aa_24082020.sh
echo "/hydra/software/interproscan-5.46-81.0/interproscan.sh -i /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/aa100longer_merged_ip.fa.pep -d /hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_more100aa_24082020 -goterms " >> interproscan_more100aa_24082020.sh
chmod +x interproscan_more100aa_24082020.sh
less interproscan_more100aa_24082020.sh

#%%bash
#cd ~/sexual_genome_annotation_files/ncrna_Neiro/
#sudo nohup ./interproscan_more100aa_24082020.sh &

### 4.3.2 Sequences shorter than 1000 amino acids

Interproscan was run on sequences shorter than 100 amino acids.

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
grep -c ">" aa100shorter_merged_ip.fa.pep

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/
sudo nohup /hydra/software/interproscan-5.46-81.0/interproscan.sh -i /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/aa100shorter_merged_ip.fa.pep -d /hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020 -goterms & 

# 4.4 Characterisation of transcripts

The new transcripts were characterised: 

### 4.4.1 Coding sequences longer than 100 amino acids

All new transcripts with an open reading frame longer than 100 amino acids were selected:

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
ls

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
grep ">" aa100longer_merged_ip.fa.pep > aa100longer_merged_ip.list.txt

import pandas as pd
aa100longer_orf_pd = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/aa100longer_merged_ip.list.txt", header=None)

aa100longer_orf = []
for i in range(0, len(aa100longer_orf_pd)):
    aa100longer_orf.append(aa100longer_orf_pd.iloc[i,0].split(" ")[0].split(">")[1].split(".p")[0])

aa100longer_orf = pd.Categorical(aa100longer_orf).categories

len(aa100longer_orf)

import pandas as pd
ncrna_list = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/ncrna_list.csv", header=None)

aa100longer_final = ncrna_list[ncrna_list.iloc[:,0].isin(aa100longer_orf)]

len(aa100longer_final)

#import pandas as pd
#pd.Series(pd.Categorical(aa100longer_final.iloc[:,0]).categories).to_csv("../sexual_genome_annotation_files/ncrna_Neiro/interproscan_more100aa/codingmore100aa.csv")

%%bash
wc -l /hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_more100aa/codingmore100aa.csv

In total, there were 2002 new transcripts with an open reading frame longer than 100 amino acids.

import pandas as pd
interpro_more100aa_df = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_more100aa_24082020/aa100longer_merged_ip.fa.pep.tsv", sep="\t",  header=None)

interpro_more100aa_genes = []
for i in range(0, len(interpro_more100aa_df)):
    interpro_more100aa_genes.append(interpro_more100aa_df.iloc[i,0].rpartition(".")[0])
interpro_more100aa_df["gene"] = interpro_more100aa_genes
interpro_more100aa_genes = pd.Categorical(interpro_more100aa_genes).categories

len(interpro_more100aa_genes)

In total, 1320 of the 2002 transcripts had an Interproscan hit and were thus classified as coding transcripts.

GO terms were assigned to the new coding sequences longer than 100 amino acids:

import pandas as pd
ncrna_list = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/ncrna_list.csv", header=None)

interpro_more100aa_GO_draft = interpro_more100aa_df[interpro_more100aa_df["gene"].isin(ncrna_list.iloc[:,0].to_list())]

interpro_more100aa_GO = interpro_more100aa_GO_draft[interpro_more100aa_GO_draft.iloc[:,13].notnull()]

interpro_more100aa_GO_genes = []
interpro_more100aa_GO_go = []
for i in range(0, len(interpro_more100aa_GO)):
    gene = interpro_more100aa_GO.iloc[i,0].rpartition(".")[0]
    gene2 = interpro_more100aa_GO.iloc[i,0]
    if gene not in interpro_more100aa_GO_genes:
        interpro_more100aa_GO_genes.append(gene)
        df = interpro_more100aa_GO[interpro_more100aa_GO.iloc[:,0] == gene2]
        gos_final = []
        for j in range(0, len(df)):
            gos = df.iloc[j, 13].split("|")
            for k in gos:
                if k not in gos_final:
                    gos_final.append(k)
        interpro_more100aa_GO_go.append(gos_final)

f = open("/hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_more100aa_24082020/interpro_more100aa.GO.May2021.tab", "a+")
for i in range(0, len(interpro_more100aa_GO_genes)):
    f.write(str(interpro_more100aa_GO_genes[i])+"\t")
    for j in range(0, len(interpro_more100aa_GO_go[i])):
        f.write(str(interpro_more100aa_GO_go[i][j])+", ")
        if j == (len(interpro_more100aa_GO_go[i])-1):
            f.write(str(interpro_more100aa_GO_go[i][j])+"\n")

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_more100aa_24082020/
wc -l interpro_more100aa.GO.May2021.tab

### 4.4.2 Coding sequences shorter than 100 amino acids

All new transcripts with an open reading frame shorter than 100 amino acids were selected:

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
grep ">" aa100shorter_merged_ip.fa.pep > aa100shorter_merged_ip.list.txt

import pandas as pd
aa100shorter_orf_pd = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/aa100shorter_merged_ip.list.txt", header=None)

aa100shorter_orf = []
for i in range(0, len(aa100shorter_orf_pd)):
    aa100shorter_orf.append(aa100shorter_orf_pd.iloc[i,0].split(" ")[0].split(">")[1].split(".p")[0])

aa100shorter_orf = pd.Categorical(aa100shorter_orf).categories

len(aa100shorter_orf)

aa100shorter_final = ncrna_list[ncrna_list.iloc[:,0].isin(aa100shorter_orf)]

import pandas as pd
pd.Series(pd.Categorical(aa100shorter_final.iloc[:,0]).categories).to_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/codingless100aa.csv")

%%bash
wc -l /hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/codingless100aa.csv

import pandas as pd
interpro_less100aa_df = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/aa100shorter_merged_ip.fa.pep.tsv", sep="\t",  header=None)

interpro_less100aa_genes = []
for i in range(0, len(interpro_less100aa_df)):
    interpro_less100aa_genes.append(interpro_less100aa_df.iloc[i,0].rpartition(".")[0])
interpro_less100aa_df["gene"] = interpro_less100aa_genes
interpro_less100aa_genes = pd.Categorical(interpro_less100aa_genes).categories

len(interpro_less100aa_genes)

interpro_less100aa_genes = interpro_less100aa_genes[~interpro_less100aa_genes.isin(interpro_more100aa_genes)]

len(interpro_less100aa_genes)

In total, 3841 transcripts of the 11252 had an interpro hit and were classified as coding, and were not present in the longer sequences already.

GO terms were assigned to the new coding sequences shorter than 100 amino acids:

interpro_less100aa_GO_draft = interpro_less100aa_df[interpro_less100aa_df["gene"].isin(ncrna_list.iloc[:,0].to_list())]

interpro_less100aa_GO = interpro_less100aa_GO_draft[interpro_less100aa_GO_draft.iloc[:,13].notnull()]

interpro_less100aa_GO = interpro_less100aa_GO[interpro_less100aa_GO["gene"].isin(interpro_less100aa_genes)]

interpro_less100aa_GO_genes = []
interpro_less100aa_GO_go = []
for i in range(0, len(interpro_less100aa_GO)):
    gene = interpro_less100aa_GO.iloc[i,0].rpartition(".")[0]
    gene2 = interpro_less100aa_GO.iloc[i,0]
    if gene not in interpro_less100aa_GO_genes:
        interpro_less100aa_GO_genes.append(gene)
        df = interpro_less100aa_GO[interpro_less100aa_GO.iloc[:,0] == gene2]
        gos_final = []
        for j in range(0, len(df)):
            gos = df.iloc[j, 13].split("|")
            for k in gos:
                if k not in gos_final:
                    gos_final.append(k)
        interpro_less100aa_GO_go.append(gos_final)

f = open("/hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/interpro_less100aa.GO.May2021.tab", "a+")
for i in range(0, len(interpro_less100aa_GO_genes)):
    f.write(str(interpro_less100aa_GO_genes[i])+"\t")
    for j in range(0, len(interpro_less100aa_GO_go[i])):
        f.write(str(interpro_less100aa_GO_go[i][j])+", ")
        if j == (len(interpro_less100aa_GO_go[i])-1):
            f.write(str(interpro_less100aa_GO_go[i][j])+"\n")

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/
wc -l interpro_less100aa.GO.May2021.tab

Merging the GO table for all new coding transcripts:

%%bash
cd /hydra/GO/interpro/
cat /hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_more100aa_24082020/interpro_more100aa.GO.May2021.tab /hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/interpro_less100aa.GO.May2021.tab > ncrna_new_coding_May2021.GO.tab

Merging the GO table for all coding transcripts:

%%bash
cd /hydra/GO/interpro/
cat /hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_more100aa_24082020/interpro_more100aa.GO.May2021.tab /hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/interpro_less100aa.GO.May2021.tab interpro_rink.GO.tab > ncrna_coding_May2021.GO.tab

%%bash
cd /hydra/GO/interpro/
wc -l ncrna_new_coding_May2021.GO.tab

### 4.4.3 Non-coding transcripts

The list of non-coding transcripts was defined as the complement of the new coding transcripts from the previous section. 

import pandas as pd
ncrna_list = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/ncrna_list.csv", header=None)

noncoding = ncrna_list[~ncrna_list.iloc[:,0].isin(interpro_more100aa_genes)]
noncoding = noncoding[~noncoding.iloc[:,0].isin(interpro_less100aa_genes)]

coding = ncrna_list[ncrna_list.iloc[:,0].isin(interpro_more100aa_genes)|ncrna_list.iloc[:,0].isin(interpro_less100aa_genes)]

The non-coding transcripts were saved as a csv file.

noncoding.to_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/noncoding.csv", index=False, header=None)

%%bash
wc -l /hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/noncoding.csv

coding.to_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/coding_new.csv", index=False, header=None)

%%bash
wc -l /hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/coding_new.csv

The non-coding transcripts with the non-codiing transcripts without ORFs:

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/
cat noncoding10aa_list.csv interproscan_less100aa_24082020/noncoding.csv > noncoding_final_May2021.csv

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/
grep -f noncoding_final_May2021.csv stringtie_merged_exp.gtf > noncoding_final_May2021.gtf

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/
grep -f interproscan_less100aa_24082020/coding_new.csv stringtie_merged_exp.gtf > coding_final_May2021.gtf

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/
wc -l noncoding_final_May2021.gtf

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/
wc -l coding_final_May2021.gtf

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/
wc -l noncoding_final_May2021.csv

The final frequencies are:
* All transcripts: 91068
* Identical to Rink: 28624
* Novel coding: 5161
* Novel non-coding: 6191

# FINNISHED
