==> notebooks/1-3-FACSexpression.ipynb <==
# 1-3-FACSexpression
Jakke Neiro$^1$
1. Aboobaker laboraotry, Department of Zoology, University of Oxford



## Contents of the notebook
* 1. Introduction
* 2. Alignment
    * 2.1 Hisat alignment
    * 2.2 Salmon pseudoalingment
* 3. Coverage
* 4. Proportional TPM values
    * 4.1 Calculation of proportional values
    * 4.2 Inspection of proportional values
* 5. Transcripts per gene

## Files
* Input:
    * oenalx1_1.fastq.gz, oenalx1_2.fastq.gz, oenalx2_1.fastq.gz, oenalx2_2.fastq.gz, oenalxins_1.fastq.gz, oenalxins_2.fastq.gz
    * sleuth_gene.csv
* Output:
    * /hydra/FACS/salmon_bootstrap
    * x1_mean_final.bw
    * FACS_prop.csv

# 1. Introduction

RNA-seq samples from X1, X2 and Xins compartments were used to generate coverage tracks and compartment-specific TPM values for X1, X2 and Xins.

# 2. Alignment

## 2.1 Hisat alignment

The Önal et al. (2012) X1, X2 and Xins RNA-seq samples were aligned to the genome. with Hisat2. The other RNA-seq samples had already been aligned in 1-1-Genomeannotation. 

%%bash
cd /hydra/neiro_bam/
less hisat2_align_oenal.sh

#%%bash
#cd /hydra/neiro_bam/
#sudo nohup ./hisat2_align.sh &

The sam files were converted to bam files.

%%bash
cd /hydra/neiro_bam/
less hisat_samtools_oenal.sh

#%%bash
#cd /hydra/neiro_bam/
#sudo nohup ./hisat_samtools_oenal.sh &

The bam files were sorted:

#%%bash
#cd /hydra/neiro_bam/
#samtools sort -@ 8 -o oenalx1.sorted.bam oenalx1.bam
#samtools sort -@ 8 -o oenalx2.sorted.bam oenalx2.bam
#samtools sort -@ 8 -o oenalxins.sorted.bam oenalxins.bam

The aligned files were later used to generate coverage tracks (bigwigs).

## 2.2 Salmon pseudoalignment

First, the transcripts were extracted by using the new annotation: 

%%bash
cd /hydra/FACS
gffread -w stringtie_transcripts.fa -g ~/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa /hydra/sexual_genome_annotation_files/ncrna_Neiro/stringtie_merged.gtf

%%bash
cd /hydra/FACS
grep -c ">" stringtie_transcripts.fa
grep ">" stringtie_transcripts.fa | head -5

The transcripts were used to generate the salmon index: 

%%bash
cd ../FACS
echo "#!/bin/bash" > salmon_index_stringtie.sh
echo "~/software/salmon/bin/salmon index -t ~/FACS/stringtie_transcripts.fa -i salmon_index_stringtie -k 31" >> salmon_index_stringtie.sh
chmod +x salmon_index_stringtie.sh
less salmon_index_stringtie.sh

#%%bash
#cd /hydra/FACS
#./salmon_index_stringtie.sh

Salmon was used to pseudoalign samples to the transcripts: 

%%bash
cd ../FACS
echo "#!/bin/bash" > salmon_quant_bootstrap.sh
echo "for f in ../../../../../data/ngs/smed/facs_rnaseq_datasets/sexual/anish_trimmed_files/*fastq.gz; do ~/software/salmon/bin/salmon quant -i salmon_index_stringtie -l A -r \$f --validateMappings -o salmon_bootstrap/\${f:77} --numBootstraps 50; done" >> salmon_quant_bootstrap.sh
chmod +x salmon_quant_bootstrap.sh
less salmon_quant_bootstrap.sh

#%%bash
#cd ../FACS
#nohup ./salmon_quant_bootstrap.sh &

#%%bash
#cd ../FACS
#nohup ~/software/salmon/bin/salmon quant -i salmon_index_stringtie -l A -1 ../../../../../data/ngs/smed/facs_rnaseq_datasets/sexual/neiro_bam/oenalx1_1.fastq.gz -2 ../../../../../data/ngs/smed/facs_rnaseq_datasets/sexual/neiro_bam/oenalx1_2.fastq.gz --validateMappings -o salmon_bootstrap/oenalx1 --numBootstraps 50 & 
#nohup ~/software/salmon/bin/salmon quant -i salmon_index_stringtie -l A -1 ../../../../../data/ngs/smed/facs_rnaseq_datasets/sexual/neiro_bam/oenalx2_1.fastq.gz -2 ../../../../../data/ngs/smed/facs_rnaseq_datasets/sexual/neiro_bam/oenalx2_2.fastq.gz --validateMappings -o salmon_bootstrap/oenalx2 --numBootstraps 50 &
#nohup ~/software/salmon/bin/salmon quant -i salmon_index_stringtie -l A -1 ../../../../../data/ngs/smed/facs_rnaseq_datasets/sexual/neiro_bam/oenalxins_1.fastq.gz -2 ../../../../../data/ngs/smed/facs_rnaseq_datasets/sexual/neiro_bam/oenalxins_2.fastq.gz --validateMappings -o salmon_bootstrap/oenalxins --numBootstraps 50 &

%%bash
cd /hydra/FACS/salmon_bootstrap
ls -l | wc -l

The names of all samples were included in a list: 

%%bash
cd /hydra/FACS/salmon_bootstrap
ls -d rna* > rna_list_tpm.txt

%%bash
cd /hydra/FACS/salmon_bootstrap
cd rna_bartscherer_tail_foxd_000h_SRR1012785_1.fastq.trimmed.fastq.gz
ls

# 3. Coverage

The bam files for X1, X2 and Xins samples were used to generate coverage tracks (bigwig files).

%%bash
cd /hydra/neiro_bam
echo "#!/bin/bash" > rna_coverage.sh
echo "cd ~/../../../../data/ngs/smed/facs_rnaseq_datasets/sexual/neiro_bam" >> rna_coverage.sh
echo "samtools index -b oenalx1.sorted.bam" >> rna_coverage.sh
echo "samtools index -b oenalx2.sorted.bam" >> rna_coverage.sh
echo "samtools index -b oenalxins.sorted.bam" >> rna_coverage.sh
echo "samtools index -b rna_pearson_x1_na_na_SRR2009674_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_pearson_x1_na_na_SRR2009675_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_pearson_x2_na_na_SRR2009676_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_pearson_x2_na_na_SRR2009677_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_reddien_x1_na_000h_SRR1302023_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_reddien_x1_na_000h_SRR1302024_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_reddien_x2_na_000h_SRR1302025_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_reddien_x2_na_000h_SRR1302026_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_sanchez_x1_na_na_SRR2407874_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_sanchez_x1_na_na_SRR2407875_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_sanchez_x1_na_na_SRR2407876_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_sanchez_x1_na_na_SRR2407877_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_sanchez_xins_na_na_SRR2407878_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_sanchez_xins_na_na_SRR2407879_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_sanchez_xins_na_na_SRR2407880_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh
echo "samtools index -b rna_sanchez_xins_na_na_SRR2407881_1.fastq.trimmed.fastq.bam" >> rna_coverage.sh

echo "bamCoverage -b oenalx1.sorted.bam -o oenalx1.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b oenalx2.sorted.bam -o oenalx2.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b oenalxins.sorted.bam -o oenalxins.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_pearson_x1_na_na_SRR2009674_1.fastq.trimmed.fastq.bam -o pearson_x1_1.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_pearson_x1_na_na_SRR2009675_1.fastq.trimmed.fastq.bam -o pearson_x1_2.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_pearson_x2_na_na_SRR2009676_1.fastq.trimmed.fastq.bam -o pearson_x2_1.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_pearson_x2_na_na_SRR2009677_1.fastq.trimmed.fastq.bam -o pearson_x2_2.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_reddien_x1_na_000h_SRR1302023_1.fastq.trimmed.fastq.bam -o reddien_x1_1.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_reddien_x1_na_000h_SRR1302024_1.fastq.trimmed.fastq.bam -o reddien_x1_2.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_reddien_x2_na_000h_SRR1302025_1.fastq.trimmed.fastq.bam -o reddien_x2_1.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_reddien_x2_na_000h_SRR1302026_1.fastq.trimmed.fastq.bam -o reddien_x2_2.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_sanchez_x1_na_na_SRR2407874_1.fastq.trimmed.fastq.bam -o sanchez_x1_1.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_sanchez_x1_na_na_SRR2407875_1.fastq.trimmed.fastq.bam -o sanchez_x1_2.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_sanchez_x1_na_na_SRR2407876_1.fastq.trimmed.fastq.bam -o sanchez_x1_3.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_sanchez_x1_na_na_SRR2407877_1.fastq.trimmed.fastq.bam -o sanchez_x1_4.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_sanchez_xins_na_na_SRR2407878_1.fastq.trimmed.fastq.bam -o sanchez_xins_1.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_sanchez_xins_na_na_SRR2407879_1.fastq.trimmed.fastq.bam -o sanchez_xins_2.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_sanchez_xins_na_na_SRR2407880_1.fastq.trimmed.fastq.bam -o sanchez_xins_3.bw --normalizeUsingRPKM" >> rna_coverage.sh
echo "bamCoverage -b rna_sanchez_xins_na_na_SRR2407881_1.fastq.trimmed.fastq.bam -o sanchez_xins_4.bw --normalizeUsingRPKM" >> rna_coverage.sh
chmod +x rna_coverage.sh

#%%bash
#cd /hydra/neiro_bam
#nohup ./coverage_rna.sh &

Wiggletools was used to generate a mean coverage track of all the samples: 

%%bash
cd /hydra/neiro_bam/
less coverage_mean_wiggle.sh

%%bash
cd /hydra/neiro_bam/
ls x1_mean_final.wig

%%bash
cd /hydra/neiro_bam
wigToBigWig x1_mean_final.wig ~/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome x1_mean_final.bw

%%bash
cd /hydra/FACS
ls *bw

# 4. Proportional TPM values 

In order to find genes enriched in the X1 compartment, proportional TPM values were assigned to the X1, X2 and Xins.

## 4.1 Calculation of proportional values

import pandas as pd
sleuth_gene = pd.read_csv("/hydra/FACS/sleuth_gene.csv")

sleuth_gene = sleuth_gene.iloc[:,2:22]

sleuth_agg = sleuth_gene.groupby(['gene']).sum()

* X1: oenalx1, SRR2009674, SRR2009675, SRR1302023, SRR1302024, SRR2407874, SRR2407875, SRR2407876, SRR2407877
* X2: oenalx2, SRR2009676, SRR2009677, SRR1302025, SRR1302026
* Xins: oenalxins, SRR2407878, SRR2407879, SRR2407880, SRR2407881

x1_agg = sleuth_agg.iloc[:,[0,3, 4, 7,8, 11, 12, 13, 14]]
x2_agg = sleuth_agg.iloc[:,[1,5, 6, 9, 10]]
xins_agg = sleuth_agg.iloc[:,[2,15, 16, 17, 18]]

Means were calculated for each gene in each compartment:

x1_agg['mean'] = x1_agg.mean(axis=1)
x2_agg['mean'] = x2_agg.mean(axis=1)
xins_agg['mean'] = xins_agg.mean(axis=1)

The proportional values were calculated by dividing the gene expression value in each compartment by the total sum.

tpm_prop = pd.DataFrame({"X1": x1_agg['mean'], "X2": x2_agg['mean'], "Xins": xins_agg["mean"]})
tpm_prop["total"] = tpm_prop.sum(axis=1)
tpm_prop["X1.prop"] = tpm_prop["X1"]/tpm_prop["total"]
tpm_prop["X2.prop"] = tpm_prop["X2"]/tpm_prop["total"]
tpm_prop["Xins.prop"] = tpm_prop["Xins"]/tpm_prop["total"]

The original Rink IDs were added to the table: 

gffcmp = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/gffcmp.stringtie_merged.gtf.tmap", sep="\t")

rink_genes = []
for i in range(0, len(tpm_prop)):
    gene = tpm_prop.index[i]
    rink_gene = gffcmp[gffcmp["qry_gene_id"] == gene].iloc[0,0]
    if rink_gene != "-":
        rink_genes.append(rink_gene)
    else:
        rink_genes.append(gene)

tpm_prop["Rink.ID"] = rink_genes

tpm_prop.to_csv("/hydra/FACS/FACS_prop.csv")

## 4.2 Inspection of proportional values

The frequencies in different enrichment categories were calculated: 

import pandas as pd
tpm_prop = pd.read_csv("/hydra/FACS/FACS_prop.csv")

tpm_prop.head()

Total number of genes: 

len(tpm_prop)

X1 enrichment category:

len(tpm_prop[tpm_prop["X1.prop"] >= 0.5])

X2 enrichment category: 

len(tpm_prop[tpm_prop["X2.prop"] >= 0.5])

Xins enrichment category: 

len(tpm_prop[tpm_prop["Xins.prop"] >= 0.5])

X1 and X2 enrichment category: 

len(tpm_prop[((tpm_prop["X1.prop"] + tpm_prop["X2.prop"]) >= 0.75) & (tpm_prop["X1.prop"] < 0.5) & (tpm_prop["X2.prop"] < 0.5)])

X1 and Xins enrichment category:

len(tpm_prop[((tpm_prop["X1.prop"] + tpm_prop["Xins.prop"]) >= 0.75) & (tpm_prop["X1.prop"] < 0.5) & (tpm_prop["Xins.prop"] < 0.5)])

X2 and Xins enrichment category: 

len(tpm_prop[((tpm_prop["X2.prop"] + tpm_prop["Xins.prop"]) >= 0.75) & (tpm_prop["X2.prop"] < 0.5) & (tpm_prop["Xins.prop"] < 0.5)])

Ubiquitous enrichment category: 

28003-4376-5284-7944-2992-1251-2054

* Ubiquitous: 4102
* X1: 4376
* X1 and X2: 2992
* X1 and Xins: 1251
* X2: 5284
* X2 and Xins: 2054
* Xins: 7944

# 5. Transcripts per gene

import pandas as pd
gffcmp = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/gffcmp.stringtie_merged.gtf.tmap", sep="\t")

import numpy as np
genes = list(pd.Categorical(gffcmp["qry_gene_id"]).categories)
counts = np.zeros(len(genes))

len(gffcmp[gffcmp.iloc[:,3] == genes[1]])

#for i in range(len(genes)):
#    counts[i] = len(gffcmp[gffcmp.iloc[:,3] == genes[i]])
#    if i % 1000 == 0:
#        print(i)

pd.DataFrame({"Genes": genes, "Counts": counts}).to_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/genetranscriptcounts.csv", index=False)

counts = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/genetranscriptcounts.csv")

counts[counts.iloc[:,1] > 100]

# FINNISHED
