==> notebooks/1-4-FACSexpression-R.ipynb <==
# 1-4-FACSexpression-R
Jakke Neiro$^1$
1. Aboobaker laboratory, Department of Zoology, University of Oxford

## Contents of notebook
* 1. Introduction
* 2. Sleuth and normalisation factors
    * 2.1 Sleuth
    * 2.2 Transcripts to genes

## Files:
* Input: salmon_bootstrap directories
* Output: sleuth_df.csv, sleuth_gene.csv

# 1. Introduction

The Salmon pseudoalignment output was used to obtain TPM values for transcripts and genes. The TPM values were used in 1-3-FACSexpression 4. Proportional TPM values. 

# 2. Sleuth and normalisation factors

## 2.1 Sleuth

Sleuth and associated packages were uploaded:

library(wasabi)
library(annotables)
library(tidyverse)
library(sleuth)

Paths to Salmon directories were specified: 

sf_dirs <- file.path("/hydra/FACS/salmon_bootstrap", 
        c("rna_pearson_x1_na_na_SRR2009674_1.fastq.trimmed.fastq.gz",
          "rna_pearson_x1_na_na_SRR2009675_1.fastq.trimmed.fastq.gz",
          "rna_pearson_x2_na_na_SRR2009676_1.fastq.trimmed.fastq.gz",
          "rna_pearson_x2_na_na_SRR2009677_1.fastq.trimmed.fastq.gz",
          "rna_reddien_x1_na_000h_SRR1302023_1.fastq.trimmed.fastq.gz",
          "rna_reddien_x1_na_000h_SRR1302024_1.fastq.trimmed.fastq.gz",
          "rna_reddien_x2_na_000h_SRR1302025_1.fastq.trimmed.fastq.gz",
          "rna_reddien_x2_na_000h_SRR1302026_1.fastq.trimmed.fastq.gz",
          "rna_sanchez_x1_na_na_SRR2407874_1.fastq.trimmed.fastq.gz",
          "rna_sanchez_x1_na_na_SRR2407875_1.fastq.trimmed.fastq.gz",
          "rna_sanchez_x1_na_na_SRR2407876_1.fastq.trimmed.fastq.gz",
          "rna_sanchez_x1_na_na_SRR2407877_1.fastq.trimmed.fastq.gz",
          "rna_sanchez_xins_na_na_SRR2407878_1.fastq.trimmed.fastq.gz",
          "rna_sanchez_xins_na_na_SRR2407879_1.fastq.trimmed.fastq.gz",
          "rna_sanchez_xins_na_na_SRR2407880_1.fastq.trimmed.fastq.gz", 
          "rna_sanchez_xins_na_na_SRR2407881_1.fastq.trimmed.fastq.gz",
          "oenalx1",
          "oenalx2",
          "oenalxins"))
sf_dirs

prepare_fish_for_sleuth(sf_dirs)

length(sf_dirs)

sf_dirs_samples = c()
for (i in 1:16){
    sf_dirs_samples = c(sf_dirs_samples, strsplit(sf_dirs[i], "_")[[1]][7])
}
sf_dirs_samples = c(sf_dirs_samples, "oenalx1", "oenalx2", "oenalxins")
sf_dirs_samples

sampletype <- factor(c(c("x1", "x1", "x2", "x2", "x1", "x1", "x2", "x2"), rep("x1", 4), rep("xins", 4), c("x1", "x2", "xins")))
sampletype

summarydata <- data.frame(sampletype, row.names = sf_dirs_samples)
all(sf_dirs_samples == rownames(summarydata))

names(sf_dirs) = sf_dirs_samples
sf_dirs

summarydata$sample <- rownames(summarydata)

summarydata$path <- sf_dirs

design <- ~ sampletype

Sleuth was used to process samples: 

so <- sleuth_prep(summarydata, full_model = design, read_bootstrap_tpm = TRUE, extra_bootstrap_summary = TRUE)

so2 = sleuth_fit(so)

TPM values were saved: 

sleuth_matrix = sleuth_to_matrix(so2, "obs_norm", "tpm")
sleuth_df = data.frame(sleuth_matrix)
write.csv(sleuth_df, "../FACS/sleuth_df.csv")

## 2.2 Transcripts to genes

sleuth_df = read.csv("/hydra/FACS/sleuth_df.csv")

The table contains transcript IDs as rows and sample names as columns:

head(sleuth_df)

nrow(sleuth_df)

gffcmp = read.table("../sexual_genome_annotation_files/ncrna_Neiro/gffcmp.stringtie_merged.gtf.tmap", head=TRUE)

head(gffcmp)

The transcripts were assigned to genes: 

sleuth_gene = sleuth_df

for (i in 1:nrow(sleuth_gene)){
    subset_gene = subset(gffcmp ,gffcmp$qry_id == toString(sleuth_gene$X[i]))
    sleuth_gene[i] = toString(subset_gene$qry_gene_id[1])
}
write.csv(sleuth_gene, "../FACS/sleuth_gene.csv")

sleuth_gene = read.csv("/hydra/FACS/sleuth_gene.csv")

# FINNISHED
