==> notebooks/1-5-BLAST.ipynb <==
# 1-6-BLAST
Jakke Neiro$^1$
1. Aboobaker laboratory, Department of Zoology, University of Oxford

## Contents of notebook
* 1. Introduction
* 2. BLAST
    * 2.1 BLAST of sequences
    * 2.2 Processing results

## Files:
* Input: stringtie_transcripts.fa, Homo_sapiens.GRCh38.pep.all.fa
* Output: human_gene_blast

# 1. Introduction

The planarian transcripts from the new annotation were blasted against human genes to find homologs.

### Conserved regions in long non-coding RNAs contain abundant translation and protein–RNA interaction signatures. 
We searched for significant sequence similarity hits between mouse annotated transcripts and human transcripts using BLASTN (45). To ensure completeness, we used a human transcriptome that contains annotated genes as well as de novo assembled transcripts obtained from high-coverage RNA sequencing data (RNA-Seq) from different tissues (46), which is available from http://dx.doi.org/10.6084/m9.figshare.4702375. The BLASTN parameters we used were: -evalue 10−5, -strand plus, -max_target_seqs 15, -window_size 12. Next, we defined ‘conserved regions’ in mouse transcripts as the ones showing significant sequence similarity (E-value < 10−5) to human transcripts, with a minimum length of 30 nucleotides. 

# 2. BLAST

## 2.1 BLAST of sequences

### 2.1.1 BLAST Coding sequences

%%bash
cd /hydra/FACS
grep -c ">" stringtie_transcripts.fa

A blast database was made for the human peptides: 

#%%bash
#cd /hydra/sexual_genome_annotation_files/blast_annotation
#makeblastdb -in Homo_sapiens.GRCh38.pep.all.fa -parse_seqids -blastdb_version 5 -dbtype prot

The planarian transcripts were blasted to the peptides: 

%%bash
cd /hydra/sexual_genome_annotation_files/blast_annotation
echo "#!/bin/bash" > blast_human.sh
echo "blastx -query /hydra/FACS/stringtie_transcripts.fa -db Homo_sapiens.GRCh38.pep.all.fa -out stringtie2human_blast.tab -outfmt 6" >> blast_human.sh
chmod +x blast_human.sh
less blast_human.sh

#%%bash
#cd /hydra/sexual_genome_annotation_files/blast_annotation
#makeblastdb -in Drosophila_melanogaster.BDGP6.32.pep.all.fa -parse_seqids -blastdb_version 5 -dbtype prot 

%%bash
cd /hydra/sexual_genome_annotation_files/blast_annotation
echo "#!/bin/bash" > blast_fly.sh
echo "blastx -query /hydra/FACS/stringtie_transcripts.fa -db Drosophila_melanogaster.BDGP6.32.pep.all.fa -out stringtie2fly_blast.tab -outfmt 6" >> blast_fly.sh
chmod +x blast_fly.sh
less blast_fly.sh

%%bash
cd /hydra/sexual_genome_annotation_files/blast_annotation
wc -l stringtie2human_blast.tab

19530460-16000000

### 2.1.2 BLAST non-coding sequences

#%%bash
#cd /drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation
#wget https://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/000/001/405/GCF_000001405.39_GRCh38.p13/GCF_000001405.39_GRCh38.p13_rna.fna.gz

#%%bash
#cd /drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation
#wget https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_38/gencode.v38.transcripts.fa.gz

#%%bash
#cd /drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation
#gunzip gencode.v38.transcripts.fa.gz

%%bash
cd /drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation
head gencode.v38.transcripts.fa

#%%bash
#cd /drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation
#makeblastdb -in GCF_000001405.39_GRCh38.p13_rna.fna -parse_seqids -blastdb_version 5 -dbtype nucl

#%%bash
#cd /drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation
#makeblastdb -in gencode_transcripts.fa -parse_seqids -blastdb_version 5 -dbtype nucl

%%bash
cd /drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation
echo "#!/bin/bash" > blast_noncodinghuman.sh
echo "blastn -query /drives/raid/AboobakerLab/jakke/FACS/noncoding_transcripts.fa -db GCF_000001405.39_GRCh38.p13_rna.fna -out noncodinghuman_blast.tab -outfmt 6" >> blast_noncodinghuman.sh
chmod +x blast_noncodinghuman.sh
less blast_noncodinghuman.sh

%%bash
cd /drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation
echo "#!/bin/bash" > blast_noncodinghuman_2.sh
echo "blastn -query /drives/raid/AboobakerLab/jakke/FACS/noncoding_transcripts.fa -db gencode_transcripts.fa -out noncodinghuman_blast_2.tab -outfmt 6" >> blast_noncodinghuman_2.sh
chmod +x blast_noncodinghuman_2.sh
less blast_noncodinghuman_2.sh

#%%bash
#cd /drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation
#nohup ./blast_noncodinghuman.sh &

#%%bash
#cd /drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation
#nohup ./blast_noncodinghuman_2.sh &

## 2.2 Processing results

### 2.2.1 Coding results

The results file were split up to make computations easier: 

%%bash
cd /hydra/sexual_genome_annotation_files/blast_annotation
head -2000000 stringtie2human_blast.tab > stringtie2human_blast_1.tab
wc -l stringtie2human_blast_1.tab
head -4000000 stringtie2human_blast.tab | tail -2000000 - > stringtie2human_blast_2.tab
wc -l stringtie2human_blast_2.tab
head -6000000 stringtie2human_blast.tab | tail -2000000 - > stringtie2human_blast_3.tab
wc -l stringtie2human_blast_3.tab
head -8000000 stringtie2human_blast.tab | tail -2000000 - > stringtie2human_blast_4.tab
wc -l stringtie2human_blast_4.tab
head -10000000 stringtie2human_blast.tab | tail -2000000 - > stringtie2human_blast_5.tab
wc -l stringtie2human_blast_5.tab
head -12000000 stringtie2human_blast.tab | tail -2000000 - > stringtie2human_blast_6.tab
wc -l stringtie2human_blast_6.tab
head -14000000 stringtie2human_blast.tab | tail -2000000 - > stringtie2human_blast_7.tab
wc -l stringtie2human_blast_7.tab
head -16000000 stringtie2human_blast.tab | tail -2000000 - > stringtie2human_blast_8.tab
wc -l stringtie2human_blast_8.tab
head -19530460 stringtie2human_blast.tab | tail -3530460 - > stringtie2human_blast_9.tab
wc -l stringtie2human_blast_9.tab

The best hit for each planarian transcript was selected: 

import pandas as pd
for i in range(1,10):
    human = pd.read_csv("/hydra/sexual_genome_annotation_files/blast_annotation/stringtie2human_blast_"+str(i)+".tab", sep="\t", header=None)
    human.columns = ["Seqid", "qid", "Per", "Len", "Mis", "Gap", "Startq", "Endq", "Starts", "Ends", "E", "Bit"]
    human = human.loc[human.groupby("Seqid")["Bit"].idxmax()]
    human.to_csv("/hydra/sexual_genome_annotation_files/blast_annotation/human"+str(i)+".tab", sep="\t", index=False, header=None)
    print(i)
    print(len(human))

%%bash
cd /hydra/sexual_genome_annotation_files/blast_annotation
cat human1.tab human2.tab human3.tab human4.tab human5.tab human6.tab human7.tab human8.tab human9.tab > human_filtered.tab

%%bash
cd /hydra/sexual_genome_annotation_files/blast_annotation
head human_filtered.tab

human = pd.read_csv("/hydra/sexual_genome_annotation_files/blast_annotation/human_filtered.tab", sep="\t", header=None)
human.columns = ["Seqid", "qid", "Per", "Len", "Mis", "Gap", "Startq", "Endq", "Starts", "Ends", "E", "Bit"]
human = human.loc[human.groupby("Seqid")["Bit"].idxmax()]
human.to_csv("/hydra/sexual_genome_annotation_files/blast_annotation/human_final.tab", sep="\t", index=False)

human = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_final.tab", sep="\t")
human[human["E"] < 10**(-10)].to_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_final_e10.tab", sep="\t", index=False)

Gene symbols and gene names were assigned to all hits: 

import pandas as pd
human = pd.read_csv("/hydra/sexual_genome_annotation_files/blast_annotation/human_final.tab", sep="\t")

import pandas as pd
mart = pd.read_csv("/hydra/sexual_genome_annotation_files/blast_annotation/mart.tsv", sep="\t")

human["Gene"] = "a"
human["Description"] = "a"
human["Synonym"] = "a"
human["Symbol"] = "a"
for i in range(len(human)):
    qid = human.iloc[i, 1].split(".")[0]
    table = mart[mart.iloc[:,4] == qid]
    human.iloc[i, 12] = table.iloc[0, 7]
    human.iloc[i, 13] = table.iloc[0, 6]
    human.iloc[i, 14] = table.iloc[0, 9]
    human.iloc[i, 15] = table.iloc[0, 10]
    if i%100 == 1:
        print(i)
human.to_csv("/hydra/sexual_genome_annotation_files/blast_annotation/human_annotated_blast.tab", sep="\t", index=False)

A table with planarian gene IDs and homologous human gene names and symbols was generated: 

import pandas as pd
human = pd.read_csv("/hydra/sexual_genome_annotation_files/blast_annotation/human_annotated_blast.tab", sep="\t")

import pandas as pd
gffcmp = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/gffcmp.stringtie_merged.gtf.tmap", sep="\t")
gene2transcript = gffcmp.iloc[:,[0,3,4]]

human["geneID"] = "a"
for i in range(len(human)):
    seqid= human.iloc[i, 0]
    human.iloc[i, 16] = gene2transcript[gene2transcript.iloc[:,2] == seqid].iloc[0,1]
    if i%1000 == 1:
        print(i)
human.to_csv("/hydra/sexual_genome_annotation_files/blast_annotation/human_gene_blast.tab", sep="\t", index=False)

import pandas as pd
human_gene_blast = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_gene_blast.tab", sep="\t")

human_gene_blast

10**(-10)

human_gene_blast[human_gene_blast["E"] < 10**(-10)].to_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_gene_blast_e10.tab", sep="\t", index=False)

### 2.2.2 Non-coding results

%%bash
cd /drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation
ls -l noncodinghuman_blast.tab 

import pandas as pd
human_non = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/noncodinghuman_blast.tab", sep="\t", header=None)
human_non.columns = ["Seqid", "qid", "Per", "Len", "Mis", "Gap", "Startq", "Endq", "Starts", "Ends", "E", "Bit"]
human_non = human_non.loc[human_non.groupby("Seqid")["Bit"].idxmax()]
human_non.to_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_noncod_final.tab", sep="\t", index=False)

import pandas as pd
human_non = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/noncodinghuman_blast.tab", sep="\t", header=None)
human_non.columns = ["Seqid", "qid", "Per", "Len", "Mis", "Gap", "Startq", "Endq", "Starts", "Ends", "E", "Bit"]
human_non = human_non.loc[human_non.groupby("Seqid")["Bit"].idxmax()]
human_non[human_non["E"] < 10**(-5)].to_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_noncod_final_e10.tab", sep="\t", index=False)

# 3. Transcriptomes of Schmidtea and Dugesia

## 3.1 BLAST to Dugesia japonica

#%%bash
#cd /hydra/Dugesia/TF_interpro
#gffread -w Dugesia_transcripts.fa -g /hydra/Dugesia/Dj_genome_v1.fasta /hydra/Dugesia/Dugesia_Oxford_v1.gtf

%%bash
cd /hydra/Dugesia/TF_interpro
ls *fa

%%bash
cd /hydra/Dugesia/TF_interpro
makeblastdb -in Dugesia_transcripts.fa -parse_seqids -dbtype nucl

%%bash
cd /hydra/Dugesia/TF_interpro
echo "#!/bin/bash" > blast_smeddjap.sh
echo "blastn -query /hydra/FACS/stringtie_transcripts.fa -db Dugesia_transcripts.fa -outfmt 6 -out smed_to_djap_trans.out -num_threads 8" >> blast_smeddjap.sh
chmod +x blast_smeddjap.sh
less blast_smeddjap.sh

%%bash
cd /hydra/FACS/
gffread -w noncoding_transcripts.fa -g /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa /hydra/sexual_genome_annotation_files/ncrna_Neiro/noncoding_final_May2021.gtf

%%bash
cd /hydra/FACS/
grep -c ">" noncoding_transcripts.fa

%%bash
cd /hydra/Dugesia/TF_interpro
echo "#!/bin/bash" > blast_smeddjap_2.sh
echo "blastn -query /hydra/FACS/noncoding_transcripts.fa -db Dugesia_transcripts.fa -outfmt 6 -out smed_to_djap_trans_param.out -num_threads 8 -word_size 6" >> blast_smeddjap_2.sh
chmod +x blast_smeddjap_2.sh
less blast_smeddjap_2.sh

%%bash
cd /hydra/FACS/
ls

%%bash
cd /hydra/FACS/
seqkit range -r 1:5000 stringtie_transcripts.fa > stringtie_transcripts_1.fa
seqkit range -r 5001:10000 stringtie_transcripts.fa > stringtie_transcripts_2.fa
seqkit range -r 10001:15000 stringtie_transcripts.fa > stringtie_transcripts_3.fa
seqkit range -r 15001:20000 stringtie_transcripts.fa > stringtie_transcripts_4.fa
seqkit range -r 20001:25000 stringtie_transcripts.fa > stringtie_transcripts_5.fa
seqkit range -r 25001:30000 stringtie_transcripts.fa > stringtie_transcripts_6.fa
seqkit range -r 30001:35000 stringtie_transcripts.fa > stringtie_transcripts_7.fa
seqkit range -r 35001:40000 stringtie_transcripts.fa > stringtie_transcripts_8.fa
seqkit range -r 40001:45000 stringtie_transcripts.fa > stringtie_transcripts_9.fa
seqkit range -r 45001:50000 stringtie_transcripts.fa > stringtie_transcripts_10.fa
seqkit range -r 50001:55000 stringtie_transcripts.fa > stringtie_transcripts_11.fa
seqkit range -r 55001:60000 stringtie_transcripts.fa > stringtie_transcripts_12.fa
seqkit range -r 60001:65000 stringtie_transcripts.fa > stringtie_transcripts_13.fa
seqkit range -r 65001:70000 stringtie_transcripts.fa > stringtie_transcripts_14.fa
seqkit range -r 70001:75000 stringtie_transcripts.fa > stringtie_transcripts_15.fa
seqkit range -r 75001:80000 stringtie_transcripts.fa > stringtie_transcripts_16.fa
seqkit range -r 80001:85000 stringtie_transcripts.fa > stringtie_transcripts_17.fa
seqkit range -r 85001:91068 stringtie_transcripts.fa > stringtie_transcripts_18.fa

%%bash
cd /hydra/Dugesia/TF_interpro
for i in {1..18}
  do 
     echo "#!/bin/bash" > tblastx_smeddjap_$i.sh
     echo "tblastx -query /hydra/FACS/stringtie_transcripts_$i.fa -db Dugesia_transcripts.fa -outfmt 6 -out smed_to_djap_tblastx_$i.out -num_threads 8" >> tblastx_smeddjap_$i.sh
     chmod +x tblastx_smeddjap_$i.sh
  done

%%bash
cd /hydra/Dugesia/TF_interpro
less tblastx_smeddjap_18.sh

#%%bash
#cd /hydra/Dugesia/TF_interpro
#nohup ./tblastx_smeddjap_1.sh &

import pandas as pd
for i in range(1,19):
    human = pd.read_csv("/hydra/Dugesia/TF_interpro/smed_to_djap_tblastx_"+str(i)+".out", sep="\t", header=None)
    human.columns = ["Seqid", "qid", "Per", "Len", "Mis", "Gap", "Startq", "Endq", "Starts", "Ends", "E", "Bit"]
    human = human.loc[human.groupby("Seqid")["Bit"].idxmax()]
    human.to_csv("/hydra/sexual_genome_annotation_files/blast_annotation/dugesiasmed"+str(i)+".tab", sep="\t", index=False, header=None)
    print(i)
    print(len(human))

import pandas as pd
tblastx_smedjap = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/dugesiasmed"+str(1)+".tab", sep="\t", header=None)
for i in range(2,19):
    tblastx_smedjap = pd.concat([tblastx_smedjap, pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/dugesiasmed"+str(i)+".tab", sep="\t", header=None)])
    print(len(tblastx_smedjap))

tblastx_smedjap.columns = ["Seqid", "qid", "Per", "Len", "Mis", "Gap", "Startq", "Endq", "Starts", "Ends", "E", "Bit"]
#tblastx_smedjap = tblastx_smedjap.loc[tblastx_smedjap.groupby("Seqid")["Bit"].idxmax()]
tblastx_smedjap.to_csv("/hydra/sexual_genome_annotation_files/blast_annotation/dugesiasmed_final.tab", sep="\t", index=False)

#E-10
tblastx_smedjap.columns = ["Seqid", "qid", "Per", "Len", "Mis", "Gap", "Startq", "Endq", "Starts", "Ends", "E", "Bit"]
tblastx_smedjap = tblastx_smedjap.loc[tblastx_smedjap.groupby("Seqid")["Bit"].idxmax()]
tblastx_smedjap = tblastx_smedjap[tblastx_smedjap["E"] < 10**(-10)]
tblastx_smedjap.to_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/dugesiasmed_final_e10.tab", sep="\t", index=False)

#E-5
tblastx_smedjap.columns = ["Seqid", "qid", "Per", "Len", "Mis", "Gap", "Startq", "Endq", "Starts", "Ends", "E", "Bit"]
tblastx_smedjap = tblastx_smedjap.loc[tblastx_smedjap.groupby("Seqid")["Bit"].idxmax()]
tblastx_smedjap = tblastx_smedjap[tblastx_smedjap["E"] < 10**(-5)]
tblastx_smedjap.to_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/dugesiasmed_final_e5.tab", sep="\t", index=False)

### 3.1.2 Protein to protein

%%bash
cd /hydra/Dugesia/TF_interpro/Transdecoder_all
head -2 longest_orfs_ip.pep

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/Transdecoder_final/
#cp longest_orfs.pep longest_orfs_ip.pep
#sed -i 's_*__g' longest_orfs_ip.pep
head -2 longest_orfs_ip.pep

%%bash
cd /hydra/Dugesia/TF_interpro/Transdecoder_all
makeblastdb -in longest_orfs_ip.pep -parse_seqids -dbtype prot

## 3.2 New transcripts and Dugesia

import pandas as pd
djap = pd.read_csv("/hydra/Dugesia/TF_interpro/smed_to_djap_noncod_param.out", sep="\t", header=None, skiprows=15869)
djap.columns = ["Seqid", "qid", "Per", "Len", "Mis", "Gap", "Startq", "Endq", "Starts", "Ends", "E", "Bit"]
djap = djap.loc[djap.groupby("Seqid")["Bit"].idxmax()]
djap.to_csv("/hydra/Dugesia/TF_interpro/smed_to_djap_noncod_param_filtered.out", index=None)

import pandas as pd
djap = pd.read_csv("/hydra/Dugesia/TF_interpro/smed_to_djap_noncod_param_filtered.out")

djap = djap[djap["E"] < 0.01]

len(pd.Categorical(djap["Seqid"]).categories)

import pandas as pd
coding_new = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/coding_new.csv", sep="\t", header=None)

coding_new 

import pandas as pd
noncoding_new = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/noncoding.csv", sep="\t", header=None)

print(len(noncoding_new))
print(len(noncoding_new[noncoding_new.iloc[:,0].isin(djap.iloc[:,0])]))

2432/6190

tblastx_smedjap = pd.read_csv("/hydra/sexual_genome_annotation_files/blast_annotation/dugesiasmed_final.tab", sep="\t")

print(len(coding_new))
print(len(coding_new[coding_new.iloc[:,0].isin(tblastx_smedjap.iloc[:,0])]))

table = coding_new[coding_new.iloc[:,0].isin(tblastx_smedjap.iloc[:,0])]
genes = []
for i in range(len(table)):
    genes.append(table.iloc[i,0].split(".")[1])
table = coding_new
genes_cod = []
for i in range(len(table)):
    name = table.iloc[i,0].split(".")
    if name[0] == "MSTRG":
        genes_cod.append(table.iloc[i,0].split(".")[1])
    else:
        genes_cod.append(table.iloc[i,0].split(".")[0])

print(len(pd.Categorical(genes_cod).categories))
print(len(pd.Categorical(genes).categories))

3126/3312

## 3.3 New transcripts to Metazoa

%%bash
cd /hydra/FACS/
gffread -w coding_transcripts.fa -g /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa /hydra/sexual_genome_annotation_files/ncrna_Neiro/coding_final_May2021.gtf

%%bash
cd /hydra/FACS/
echo "#!/bin/bash" > blast_smeddjap_newmetazoa.sh
echo "tblastx -query /hydra/FACS/coding_transcripts.fa -db nr -remote -outfmt 6 -out smed_newmetazoa.out -entrez_query \"Metazoa [Organism]\"" >> blast_smeddjap_newmetazoa.sh
chmod +x blast_smeddjap_newmetazoa.sh
less blast_smeddjap_newmetazoa.sh

%%bash
cd /hydra/FACS/
echo "#!/bin/bash" > blastx_smeddjap_newmetazoa.sh
echo "blastx -query /hydra/FACS/coding_transcripts.fa -db nr -remote -outfmt \"6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore sscinames scomnames\" -out smed_newmetazoa_blastx.out -entrez_query \"Metazoa [Organism] NOT Schmidtea mediterranea [Organism]\"" >> blastx_smeddjap_newmetazoa.sh
chmod +x blastx_smeddjap_newmetazoa.sh
less blastx_smeddjap_newmetazoa.sh

%%bash
cd /hydra/FACS/
grep -c ">" coding_transcripts.fa

%%bash
cd /hydra/FACS/
seqkit range -r 1:2000 coding_transcripts.fa > coding_transcripts_1.fa
seqkit range -r 2001:4000 coding_transcripts.fa > coding_transcripts_2.fa
seqkit range -r 4001:6000 coding_transcripts.fa > coding_transcripts_3.fa
seqkit range -r 6001:8000 coding_transcripts.fa > coding_transcripts_4.fa
seqkit range -r 8001:10847 coding_transcripts.fa > coding_transcripts_5.fa

%%bash
cd /hydra/FACS/
echo "#!/bin/bash" > blastx_smeddjap_newmetazoa_parts.sh
echo "blastx -query /hydra/FACS/coding_transcripts_1.fa -db nr -remote -outfmt \"6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore sscinames scomnames\" -out smed_newmetazoa_blastx1.out -entrez_query \"Metazoa [Organism] NOT Schmidtea mediterranea [Organism]\"" >> blastx_smeddjap_newmetazoa_parts.sh
echo "blastx -query /hydra/FACS/coding_transcripts_2.fa -db nr -remote -outfmt \"6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore sscinames scomnames\" -out smed_newmetazoa_blastx2.out -entrez_query \"Metazoa [Organism] NOT Schmidtea mediterranea [Organism]\"" >> blastx_smeddjap_newmetazoa_parts.sh
echo "blastx -query /hydra/FACS/coding_transcripts_3.fa -db nr -remote -outfmt \"6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore sscinames scomnames\" -out smed_newmetazoa_blastx3.out -entrez_query \"Metazoa [Organism] NOT Schmidtea mediterranea [Organism]\"" >> blastx_smeddjap_newmetazoa_parts.sh
echo "blastx -query /hydra/FACS/coding_transcripts_4.fa -db nr -remote -outfmt \"6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore sscinames scomnames\" -out smed_newmetazoa_blastx4.out -entrez_query \"Metazoa [Organism] NOT Schmidtea mediterranea [Organism]\"" >> blastx_smeddjap_newmetazoa_parts.sh
echo "blastx -query /hydra/FACS/coding_transcripts_5.fa -db nr -remote -outfmt \"6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore sscinames scomnames\" -out smed_newmetazoa_blastx5.out -entrez_query \"Metazoa [Organism] NOT Schmidtea mediterranea [Organism]\"" >> blastx_smeddjap_newmetazoa_parts.sh
chmod +x blastx_smeddjap_newmetazoa_parts.sh
less blastx_smeddjap_newmetazoa_parts.sh

import pandas as pd
human = pd.read_csv("/hydra/sexual_genome_annotation_files/blast_annotation/human_final.tab", sep="\t")

coding = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/coding_new.csv", header=None)

len(coding[coding.iloc[:,0].isin(human["Seqid"])])

len(coding)

import pandas as pd
noncoding_new = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/noncoding.csv", sep="\t", header=None)

len(noncoding_new[noncoding_new.iloc[:,0].isin(human["Seqid"])])

len(noncoding_new)

len(noncoding_new[noncoding_new.iloc[:,0].isin(human["Seqid"])&noncoding_new.iloc[:,0].isin(tblastx_smedjap["Seqid"])])

djap

len(coding[coding.iloc[:,0].isin(tblastx_smedjap["Seqid"])])

len(coding[coding.iloc[:,0].isin(human["Seqid"])&coding.iloc[:,0].isin(tblastx_smedjap["Seqid"])])

len(coding[coding.iloc[:,0].isin(human["Seqid"])&~coding.iloc[:,0].isin(tblastx_smedjap["Seqid"])])

### Gene numbers coding

import pandas as pd
#human = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_final.tab", sep="\t")
human = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_final_e10.tab", sep="\t")
#human_noncod = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_noncod_final.tab", sep="\t")
human_noncod = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_noncod_final.tab", sep="\t")
coding = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/coding_new.csv", header=None)
noncoding_new = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/noncoding.csv", sep="\t", header=None)
gffcmp = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/ncrna_Neiro/gffcmp.stringtie_merged.gtf.tmap", sep="\t")
#tblastx_smedjap = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/dugesiasmed_final.tab", sep="\t")
tblastx_smedjap = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/dugesiasmed_final_e10.tab", sep="\t")

coding_genes = []
for i in range(0, len(coding)):
    tr_id = coding.iloc[i,0]
    gene_id = gffcmp[gffcmp["qry_id"] == tr_id]["qry_gene_id"].iloc[0]
    if gene_id not in coding_genes:
        if gene_id not in list(pd.Categorical(gffcmp[~gffcmp["class_code"].isin(['u', 'i', "x"])]["qry_gene_id"]).categories):
            coding_genes.append(gene_id)

print("All new coding genes: " + str(len(coding_genes)))

print("All new coding genes e-10: " + str(len(coding_genes)))

coding_in_djaphuman = coding[coding.iloc[:,0].isin(human["Seqid"])|coding.iloc[:,0].isin(tblastx_smedjap["Seqid"])]
djaphumancoding_genes = []
for i in range(0, len(coding_in_djaphuman)):
    tr_id = coding_in_djaphuman.iloc[i,0]
    gene_id = gffcmp[gffcmp["qry_id"] == tr_id]["qry_gene_id"].iloc[0]
    if gene_id not in djaphumancoding_genes:
        if gene_id not in list(pd.Categorical(gffcmp[~gffcmp["class_code"].isin(['u', 'i', "x"])]["qry_gene_id"]).categories):
            djaphumancoding_genes.append(gene_id)

print("Coding genes homolog in Human or Djap: " + str(len(djaphumancoding_genes)))

print("Coding genes homolog in Human or Djap e-10: " + str(len(djaphumancoding_genes)))

coding_in_human = coding[coding.iloc[:,0].isin(human["Seqid"])]
humancoding_genes = []
for i in range(0, len(coding_in_human)):
    tr_id = coding_in_human.iloc[i,0]
    gene_id = gffcmp[gffcmp["qry_id"] == tr_id]["qry_gene_id"].iloc[0]
    if gene_id not in humancoding_genes:
        if gene_id not in list(pd.Categorical(gffcmp[~gffcmp["class_code"].isin(['u', 'i', "x"])]["qry_gene_id"]).categories):
            humancoding_genes.append(gene_id)

counter = 0
for i in humancoding_genes:
    if i in djaphumancoding_genes:
        counter = counter + 1
print("Human genes: " + str(counter))

counter = 0
for i in humancoding_genes:
    if i in djaphumancoding_genes:
        counter = counter + 1
print("Human genes e-10: " + str(counter))

### Gene numbers non-coding

import pandas as pd
#human = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_final.tab", sep="\t")
human = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_final_e10.tab", sep="\t")
#human_noncod = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_noncod_final.tab", sep="\t")
human_noncod = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/human_noncod_final.tab", sep="\t")
coding = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/coding_new.csv", header=None)
noncoding_new = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/ncrna_Neiro/interproscan_less100aa_24082020/noncoding.csv", sep="\t", header=None)
gffcmp = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/ncrna_Neiro/gffcmp.stringtie_merged.gtf.tmap", sep="\t")
#tblastx_smedjap = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/dugesiasmed_final.tab", sep="\t")
tblastx_smedjap = pd.read_csv("/drives/raid/AboobakerLab/jakke/sexual_genome_annotation_files/blast_annotation/dugesiasmed_final_e5.tab", sep="\t")

noncoding_genes = []
for i in range(0, len(noncoding_new)):
    tr_id = noncoding_new.iloc[i,0]
    gene_id = gffcmp[gffcmp["qry_id"] == tr_id]["qry_gene_id"].iloc[0]
    if gene_id not in noncoding_genes:
        if gene_id not in list(pd.Categorical(gffcmp[~gffcmp["class_code"].isin(['u', 'i', "x"])]["qry_gene_id"]).categories):
            noncoding_genes.append(gene_id)

print("All non-coding: " + str(len(noncoding_genes)))

print("All non-coding e5: " + str(len(noncoding_genes)))

noncoding_in_djaphuman = noncoding_new[noncoding_new.iloc[:,0].isin(human_noncod["Seqid"])|noncoding_new.iloc[:,0].isin(tblastx_smedjap["Seqid"])]
djaphumannoncoding_genes = []
for i in range(0, len(noncoding_in_djaphuman)):
    tr_id = noncoding_in_djaphuman.iloc[i,0]
    gene_id = gffcmp[gffcmp["qry_id"] == tr_id]["qry_gene_id"].iloc[0]
    if gene_id not in djaphumannoncoding_genes:
        if gene_id not in list(pd.Categorical(gffcmp[~gffcmp["class_code"].isin(['u', 'i', "x"])]["qry_gene_id"]).categories):
            djaphumannoncoding_genes.append(gene_id)

print("Non-coding genes in Human and Djap: " + str(len(djaphumannoncoding_genes)))

print("Non-coding genes in Human and Djap e5: " + str(len(djaphumannoncoding_genes)))

noncoding_in_human = noncoding_new[noncoding_new.iloc[:,0].isin(human_noncod["Seqid"])]
humannoncoding_genes = []
for i in range(0, len(noncoding_in_human)):
    tr_id = noncoding_in_human.iloc[i,0]
    gene_id = gffcmp[gffcmp["qry_id"] == tr_id]["qry_gene_id"].iloc[0]
    if gene_id not in humannoncoding_genes:
        humannoncoding_genes.append(gene_id)

counter = 0
for i in humannoncoding_genes:
    if i in djaphumannoncoding_genes:
        counter = counter + 1
print("Human non-coding loci: " + str(counter))

counter = 0
for i in humannoncoding_genes:
    if i in djaphumannoncoding_genes:
        counter = counter + 1
print("Human non-coding loci e5: " + str(counter))

3121+4291

len(coding)

len(noncoding_new)

5065+6190

coding

sum(gffcmp[gffcmp["qry_gene_id"].isin(coding_genes)]["qry_id"].isin(coding.iloc[:,0]))

sum(gffcmp[gffcmp["qry_gene_id"].isin(noncoding_genes)]["qry_id"].isin(noncoding_new.iloc[:,0]))

4752+5884

# 4. Isoform number and BLAST

counts = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/genetranscriptcounts.csv")

import pandas as pd
human_gene_blast = pd.read_csv("/hydra/sexual_genome_annotation_files/blast_annotation/human_gene_blast.tab", sep="\t")

counts["Description"] = " "
counts["Synonym"] = " "
counts["Symbol"] = " "

#for i in range(len(counts)):
#    table = human_gene_blast[human_gene_blast["geneID"] == counts.iloc[i,0]]
#    if len(table) > 0:
#        counts["Description"].iloc[i] = table["Description"].iloc[0]
#        counts["Synonym"].iloc[i] = table["Synonym"].iloc[0]
#        counts["Symbol"].iloc[i] = table["Symbol"].iloc[0]
#    if i%1000 == 1:
#        print(i)

counts.to_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/counts_annotated.csv", index=False)

import pandas as pd
counts = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/counts_annotated.csv")

counts[counts["Counts"] > 50].sort_values(by="Counts", ascending=False)

gffcmp = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/gffcmp.stringtie_merged.gtf.tmap", sep="\t")

# FINNISHED
