==> notebooks/2-1-Transcription-factors-identification.ipynb <==
# 2-1-Transcription-factors-identification
Jakke Neiro$^1$
1. Aboobaker laboratory, Department of Zoology, University of Oxford

This notebook contains code to the following publication. This notebook (2-1-Transcription-factors-identification) describes how the transcription factors were identified from the new genome annotation and how the list of transcription factors was manually curated. 

### Identification of enhancer-like elements defines regulatory networks active in planarian adult stem cells

Jakke Neiro, Divya Sridhar,  Anish Dattani, Aziz Aboobaker

https://www.biorxiv.org/content/10.1101/2022.02.03.479047v2.full

## Contents of notebook

* 1. Introduction
* 2.  Classification of protein families: interproscan
    * 2.1 Transdecoder
    * 2.2 Interproscan
* 3. Identification of transcription factors
    * 3.1 Transcription factor IDs
    * 3.2 Transcription factor IDs in interproscan result
    * 3.3 From Interproscan result to gene IDs
* 4. Manual curation

## Files
* Input: stringtie_transcripts.fa
* Output:

# 1. Introduction

Transcription factors were identified based on the new annotation. ORFs and peptide sequences were generated with Transdecoder and TF-like protein domains were scanned with Interproscan. Transcription factors were identified based on three criteria: pfam IDs, Superfamily IDs, and the description "Transcription factor" in the description. Lastly, the computationally derived TFs were manually curated. 

# 2. Classification of protein families: interproscan

## 2.1 Transdecoder

The transcripts generated from the new annotation were used (91068 transcripts in the fasta file).

%%bash
cd /hydra/FACS/
grep -c ">" stringtie_transcripts.fa

Transdecoder was used to obtain reading frames

#%%bash
#cd /hydra/FACS/
#nohup TransDecoder.LongOrfs -t stringtie_transcripts.fa -O Transdecoder_all &

The last * from each protein sequence was removed:

#%%bash
#cd /hydra/FACS/Transdecoder_all
#cp longest_orfs.pep longest_orfs_ip.pep
#sed -i 's_*__g' longest_orfs_ip.pep
#head longest_orfs_ip.pep

%%bash
cd /hydra/FACS/Transdecoder_all
grep -c ">" longest_orfs_ip.pep

#longest_orfs_ip.pep was used for Interproscan

## 2.2 Interproscan

Interproscan was performed on the sequences, including -goterms:

%%bash
cd /hydra/FACS
cd TF_interpro
echo "#!/bin/bash" > interproscan_TF.sh
echo "/hydra/software/interproscan-5.46-81.0/interproscan.sh -i /hydra/FACS/Transdecoder_all/longest_orfs_ip.pep -d /hydra/FACS/TF_interpro/interproscan_24082020 -goterms " >> interproscan_TF.sh
chmod +x interproscan_TF.sh
less interproscan_TF.sh

#%%bash
#cd /hydra/FACS/TF_interpro
#nohup ./interproscan_TF.sh &

# 3. Identification of transcription factors

Transcription factors were identified based on three criteria: pfam IDs, Superfamily IDs, and the description "Transcription factor" in the description gien by Interproscan.

## 3.1 Transcription factor IDs

The Pfam IDs were reorganised into a list:

import pandas as pd
pfam_tf = pd.read_csv("/hydra/sexual_genome_annotation_files/Swapna/pfam_tf.txt", sep="\t", header=None)

pfam_tf.iloc[:,1].to_csv("/hydra/sexual_genome_annotation_files/Swapna/pfam_tf_list.txt", index=False, header=None)

The Superfamily IDs were reorganised into a list: 

import pandas as pd
superfamily_tf = pd.read_csv("/hydra/sexual_genome_annotation_files/Swapna/SUPERFAMILY_1_69.dbds.v2.03.txt", sep="\t", header=None)
superfamily_tf_id = []
for i in range(0, len(superfamily_tf)):
    superfamily_tf_id.append("SSF"+str(superfamily_tf.iloc[i,5]))
superfamily_tf_id[0:5]

pd.DataFrame(superfamily_tf_id).to_csv("/hydra/sexual_genome_annotation_files/Swapna/superfamily_list.txt", index=False, header=None)

%%bash
cd /hydra/sexual_genome_annotation_files/Swapna
head -2 superfamily_list.txt

#The pfam_tf_list.txt and superfamily_list.txt were used to select transcription factors from the Interproscan result. 

## 3.2 Transcription factor IDs in interproscan result

%%bash
cd /hydra/FACS/TF_interpro/interproscan_24082020
wc -l longest_orfs_ip.pep.tsv

Transcription factors were chosen based on pfam_ids:

#%%bash
#cd /hydra/FACS/TF_interpro/interproscan_24082020
#while read p; do grep -w $p longest_orfs_ip.pep.tsv >> pfam.tf.pep.tsv ; done < /hydra/sexual_genome_annotation_files/Swapna/pfam_tf_list.txt

%%bash
cd /hydra/FACS/TF_interpro/interproscan_24082020
wc -l pfam.tf.pep.tsv

Transcription factors were chosen based on superfamily_ids:

#%%bash
#cd /hydra/FACS/TF_interpro/interproscan_24082020
#while read p; do grep -w $p longest_orfs_ip.pep.tsv >> superfamily.tf.pep.tsv ; done < /hydra/sexual_genome_annotation_files/Swapna/superfamily_list.txt

%%bash
cd /hydra/FACS/TF_interpro/interproscan_24082020
wc -l superfamily.tf.pep.tsv

Transcription factors were chosen based on the name "Transcription factor" in Interproscan results:

%%bash
cd /hydra/FACS/TF_interpro/interproscan_24082020
grep "Transcription factor" longest_orfs_ip.pep.tsv > transcription_factor.tf.pep.tsv

%%bash
cd /hydra/FACS/TF_interpro/interproscan_24082020
wc -l transcription_factor.tf.pep.tsv

The results were combined into a single file:

%%bash
cd /hydra/FACS/TF_interpro/interproscan_24082020
cat pfam.tf.pep.tsv superfamily.tf.pep.tsv transcription_factor.tf.pep.tsv > final.tf.pep.tsv

%%bash
cd /hydra/FACS/TF_interpro/interproscan_24082020
wc -l final.tf.pep.tsv

## 3.3 From Interproscan result to gene IDs 

import pandas as pd
interpro_all_df = pd.read_csv("/hydra/FACS/TF_interpro/interproscan_24082020/final.tf.pep.tsv", sep="\t", header=None, low_memory=False)

len(interpro_all_df)

Protein IDs were converted into transcript ids:

interpro_all_df_transcript_ids = []
for i in range(len(interpro_all_df)):
    id_i = interpro_all_df.iloc[i,0].split(".p")[0]
    if id_i not in interpro_all_df_transcript_ids:
        interpro_all_df_transcript_ids.append(id_i)

The table containing the correspondence between transcript ID and gene ID was uploaded:

gffcmp = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/gffcmp.stringtie_merged.gtf.tmap", sep="\t")
gene2transcript = gffcmp.iloc[:,[0,3,4]]

Transcripts with a TF ID were selected:

tf_transcripts = gene2transcript[gene2transcript["qry_id"].isin(interpro_all_df_transcript_ids)]

qry_gene_id = pd.Categorical(tf_transcripts["qry_gene_id"]).categories.to_list()
ref_gene_id = []
for i in range(len(qry_gene_id)):
    ref_gene_id.append(tf_transcripts[tf_transcripts["qry_gene_id"] == qry_gene_id[i]].iloc[0,0])
tf_genes_Aug2020 = pd.DataFrame({"Rink": ref_gene_id, "Neiro": qry_gene_id})

The table is saved as a csv file.

tf_genes_Aug2020.to_csv("/hydra/FACS/TF_interpro/tf_genes_Aug2020.csv", index=0)

The IDs are saved as a txt file.

tf_genes_Aug2020["Neiro"].to_csv("/hydra/FACS/TF_interpro/tf_genes_Aug2020_NeiroIDs.txt", index=0, header=None)

This computationally predicted list of putative transcription factors was manually curated.

# 4. Manual curation

The results were manually curated and transcripts blasting to transposons or non-TF genes were removed. The resulting file was tf_genes_April2021.csv.

The planarian literature was reviewed and TFs that had not been computationally derived were added to the list. The resulting file tf_genes_May2021.xlsx contains the literature-derived TFs.

%%bash
ls -l /hydra/TF_data/tf_genes_April2021.csv
ls -l /hydra/TF_data/tf_genes_May2021.xlsx

import pandas as pd
pd.read_excel("/hydra/TF_data/tf_genes_May2021.xlsx").to_csv("/hydra/TF_data/tf_genes_May2021.xlsx", index=False)

The review of planarian literature is described in 2-2-Transcription-factors-literature-2000-2009, 

# FINNISHED
