==> notebooks/2-5-Transcription-factors-motif-inference.ipynb <==
# 2-5-Transcription-factors-motif-inference
Jakke Neiro$^1$
1. Aboobaker laboratory, Department of Zoology, University of Oxford

## Contents of notebook
* 1. Introduction
* 2. Transcription factor sequences
    * 2.1 Selection of sequences
    * 2.2 Protein sequences
* 3. Motif inference
    * 3.1 Jaspar profile inference
    * 3.2 Jaspar result processing
    * 3.3 Downloading motifs
* 4. Final transcription factor table

## Files
* Input: tf_genes_April2021.csv
* Output: all.meme

# 1. Introduction

Motifs were predicted for all transcription factors using JASPAR.

# 2. Transcription factor sequences

## 2.1 Selection of sequences

Firstly, the computationally predicted and manually curated sequences were used:

%%bash
cd /hydra/TF_data
ls tf_genes_April2021.csv 

import pandas as pd
tf_genes = pd.read_csv("/hydra/TF_data/tf_genes_April2021.csv", sep=";")

gffcmp = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/gffcmp.stringtie_merged.gtf.tmap", sep="\t")
gene2transcript = gffcmp.iloc[:,[0,3,4]]

The IDs were saved as a list: 

gene2transcript[gene2transcript["qry_gene_id"].isin(tf_genes.iloc[:,1])].iloc[:,2].to_csv("/hydra/TF_data/tf_genes_April2021.list", index=False, header=False)

%%bash
cd /hydra/TF_data
head tf_genes_April2021.list

Sequences were extracted as a fasta file:

%%bash
cd /hydra/TF_data
seqtk subseq /hydra/FACS/stringtie_transcripts.fa tf_genes_April2021.list > tf_genes_April2021.fasta

%%bash
cd /hydra/TF_data
grep ">" tf_genes_April2021.fasta | head -2
head -2 tf_genes_April2021.fasta

Secondly, the sequences from literature were used:

%%bash
cd /hydra/TF_data
ls tf_genes_May2021.xlsx

import pandas as pd
tf_genes_lit = pd.read_excel("/hydra/TF_data/tf_genes_May2021.xlsx")

tf_genes_lit.iloc[565:,:]

import pandas as pd
gffcmp = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/gffcmp.stringtie_merged.gtf.tmap", sep="\t")
gene2transcript = gffcmp.iloc[:,[0,3,4]]

The IDs were saved as a list: 

gene2transcript[gene2transcript["qry_gene_id"].isin(tf_genes_lit.iloc[565:,1])].iloc[:,2].to_csv("/hydra/TF_data/tf_genes_May2021lit.list", index=False, header=False)

%%bash
cd /hydra/TF_data
head -2 tf_genes_May2021lit.list

Sequences were extracted as a fasta file:

%%bash
cd /hydra/TF_data
seqtk subseq /hydra/FACS/stringtie_transcripts.fa tf_genes_May2021lit.list > tf_genes_May2021lit.fasta

%%bash
cd /hydra/TF_data
grep -c ">" tf_genes_May2021lit.fasta 

## 2.2 Protein sequences

Firstly, the ORFs and protein sequences of the computationally predicted TFs were extracted:

#%%bash
#cd /hydra/TF_data
#nohup TransDecoder.LongOrfs -t tf_genes_April2021.fasta -O Transdecoder_tfApril2021 &

#%%bash
#cd /hydra/TF_data/Transdecoder_tfApril2021
#cp longest_orfs.pep longest_orfs_ip.pep
#sed -i 's_*__g' longest_orfs_ip.pep
#head longest_orfs_ip.pep

%%bash
cd /hydra/TF_data/Transdecoder_tfApril2021
awk 'NR % 2 == 0' longest_orfs_ip.pep > longest_orfs_ip_seq.pep
awk 'NR % 2 == 1' longest_orfs_ip.pep > longest_orfs_ip_id.pep

%%bash
cd /hydra/TF_data/Transdecoder_tfApril2021
head -1 longest_orfs_ip_seq.pep
head -1 longest_orfs_ip_id.pep

%%bash
cd /hydra/TF_data/Transdecoder_tfApril2021
wc -l longest_orfs_ip_seq.pep

Secondly, the ORFs and protein sequences of TFs derived from literature were extracted: 

#%%bash
#cd /hydra/TF_data
#TransDecoder.LongOrfs -t tf_genes_May2021lit.fasta -O Transdecoder_tfMay2021lit

#%%bash
#cd /hydra/TF_data/Transdecoder_tfMay2021lit
#cp longest_orfs.pep longest_orfs_ip.pep
#sed -i 's_*__g' longest_orfs_ip.pep
#head longest_orfs_ip.pep

%%bash
cd /hydra/TF_data/Transdecoder_tfMay2021lit
awk 'NR % 2 == 0' longest_orfs_ip.pep > longest_orfs_ip_seq.pep
awk 'NR % 2 == 1' longest_orfs_ip.pep > longest_orfs_ip_id.pep

%%bash
cd /hydra/TF_data/Transdecoder_tfMay2021lit
head -1 longest_orfs_ip_seq.pep
head -1 longest_orfs_ip_id.pep

%%bash
cd /hydra/TF_data/Transdecoder_tfMay2021lit
wc -l longest_orfs_ip_seq.pep

# 3. Motif inference

Motifs were predcted based on the protein sequence: 

## 3.1 Jaspar profile inference

Firstly, motifs were predicted for the computationally derived TFs: 

%%bash
cd /hydra/TF_data
echo "#!/bin/bash" > jaspar_motif.sh
echo "while read p; do coreapi action infer read -p sequence=\$p >> jaspar_output.txt; done < Transdecoder_tfApril2021/longest_orfs_ip_seq.pep" >> jaspar_motif.sh
chmod +x jaspar_motif.sh
less jaspar_motif.sh

#%%bash
#cd /hydra/TF_data
#nohup ./jaspar_motif.sh &

Secondly, motifs were predicted for the TFs derived from literature:

%%bash
cd /hydra/TF_data
echo "#!/bin/bash" > jaspar_motif_lit.sh
echo "while read p; do coreapi action infer read -p sequence=\$p >> jaspar_output_lit.txt; done < Transdecoder_tfMay2021lit/longest_orfs_ip_seq.pep" >> jaspar_motif_lit.sh
chmod +x jaspar_motif_lit.sh
less jaspar_motif_lit.sh

#%%bash
#cd /hydra/TF_data
#nohup ./jaspar_motif_lit.sh &

## 3.2 Jaspar result processing

The profile inference result for the computationally derived TFs was processed into table format:

file1 = open('/hydra/TF_data/jaspar_output.txt', 'r')
Lines = file1.readlines()

import pandas as pd
profiles = pd.read_csv("/hydra/TF_data/Transdecoder_tfApril2021/longest_orfs_ip_id.pep", header=None)

profiles["TranscriptID"] = " "
profiles["GeneID"] = " "
profiles["MatrixID"] = " "
profiles["Dbd"] = 0
profiles["Evalue"] = 0

for i in range(len(profiles)):
    trans_id = profiles.iloc[i,0].split(">")[1].split(".p")[0]
    profiles.iloc[i, 1] = trans_id
    profiles.iloc[i, 2] = gene2transcript[gene2transcript["qry_id"] ==  trans_id].iloc[0,1]

import numpy as np
other_index = 0
write = 0
for i in range(len(Lines)):
    if len(Lines[i].split(" ")) > 1:
        if Lines[i].split(" ")[4] == '"count":':
            results_n = int(Lines[i].split(":")[1].split(" ")[1].split(",")[0])
            if results_n > 0:
                w = i + 6
                ma_id = Lines[i+6].split(":")[1].split(" ")[1].split(",")[0].split("\"")[1]
                dbd = Lines[i+7].split(":")[1].split(" ")[1].split(",")[0]
                evalue = Lines[i+5].split(":")[1].split(" ")[1].split(",")[0]
                profiles.iloc[other_index,3] = ma_id
                profiles.iloc[other_index,4] = float(dbd)
                profiles.iloc[other_index,5] = float(evalue)
                other_index += 1
            else:
                profiles.iloc[other_index,3] = "-"
                profiles.iloc[other_index,4] = 0.0
                profiles.iloc[other_index,5] = 1.0 
                other_index += 1

profiles_filtered = profiles.loc[profiles.groupby("GeneID")["Evalue"].idxmin()]

profiles_filtered = profiles_filtered[profiles_filtered["Evalue"] < 1]

tf_genes = pd.read_csv("/hydra/TF_data/tf_genes_April2021.csv", sep=";")

profiles_filtered["Symbol"] = " "
for i in range(len(profiles_filtered)):
    name = tf_genes[tf_genes["Neiro"] == profiles_filtered.iloc[i, 2]].iloc[0, 3]
    profiles_filtered.iloc[i, 6] = name

profiles_filtered["URL"] = " "
for i in range(len(profiles_filtered)):
    profiles_filtered.iloc[i,7] = "http://jaspar.genereg.net/api/v1/matrix/" + profiles_filtered.iloc[i,3] + "/?format=meme"

profiles_filtered.iloc[:,7].to_csv("/hydra/TF_data/April2021motifs/matrixURL.list", index=False, header=False)
profiles_filtered.iloc[:,2].to_csv("/hydra/TF_data/April2021motifs/matrixURLname.list", index=False, header=False)

profiles_filtered.to_csv("/hydra/TF_data/jaspar_filtered.csv", index=False)

The profile inference result for the TFs derived from the literature was processed into table format:

file1 = open('/hydra/TF_data/jaspar_output_lit.txt', 'r')
Lines = file1.readlines()
import pandas as pd
profiles = pd.read_csv("/hydra/TF_data/Transdecoder_tfMay2021lit/longest_orfs_ip_id.pep", header=None)
profiles["TranscriptID"] = " "
profiles["GeneID"] = " "
profiles["MatrixID"] = " "
profiles["Dbd"] = 0
profiles["Evalue"] = 0
for i in range(len(profiles)):
    trans_id = profiles.iloc[i,0].split(">")[1].split(".p")[0]
    profiles.iloc[i, 1] = trans_id
    profiles.iloc[i, 2] = gene2transcript[gene2transcript["qry_id"] ==  trans_id].iloc[0,1]
import numpy as np
other_index = 0
write = 0
for i in range(len(Lines)):
    if len(Lines[i].split(" ")) > 1:
        if Lines[i].split(" ")[4] == '"count":':
            results_n = int(Lines[i].split(":")[1].split(" ")[1].split(",")[0])
            if results_n > 0:
                w = i + 6
                ma_id = Lines[i+6].split(":")[1].split(" ")[1].split(",")[0].split("\"")[1]
                dbd = Lines[i+7].split(":")[1].split(" ")[1].split(",")[0]
                evalue = Lines[i+5].split(":")[1].split(" ")[1].split(",")[0]
                profiles.iloc[other_index,3] = ma_id
                profiles.iloc[other_index,4] = float(dbd)
                profiles.iloc[other_index,5] = float(evalue)
                other_index += 1
            else:
                profiles.iloc[other_index,3] = "-"
                profiles.iloc[other_index,4] = 0.0
                profiles.iloc[other_index,5] = 1.0 
                other_index += 1

profiles_filtered = profiles.loc[profiles.groupby("GeneID")["Evalue"].idxmin()]
profiles_filtered = profiles_filtered[profiles_filtered["Evalue"] < 1]

import pandas as pd
tf_genes_lit = pd.read_excel("/hydra/TF_data/tf_genes_May2021.xlsx")
tf_genes = tf_genes_lit.iloc[565:,:]

profiles_filtered["Symbol"] = " "
for i in range(len(profiles_filtered)):
    name = tf_genes[tf_genes["Column2"] == profiles_filtered.iloc[i, 2]].iloc[0, 3]
    profiles_filtered.iloc[i, 6] = name

profiles_filtered["URL"] = " "
for i in range(len(profiles_filtered)):
    profiles_filtered.iloc[i,7] = "http://jaspar.genereg.net/api/v1/matrix/" + profiles_filtered.iloc[i,3] + "/?format=meme"

profiles_filtered.iloc[:,7].to_csv("/hydra/TF_data/April2021motifs/matrixURLlit.list", index=False, header=False)
profiles_filtered.iloc[:,2].to_csv("/hydra/TF_data/April2021motifs/matrixURLnamelit.list", index=False, header=False)

profiles_filtered.to_csv("/hydra/TF_data/jaspar_filtered_lit.csv", index=False)

## 3.3 Downloading motifs

Based on the matrix IDs, the memes were dowbnloaded in MEME format. 

%%bash
cd /hydra/TF_data/April2021motifs
echo "#!/bin/bash" > jasparurl.sh
echo "while read p; do wget \$p -O \${p:40:6}.meme; done < matrixURL.list" >> jasparurl.sh
chmod +x jasparurl.sh
less jasparurl.sh

#%%bash
#cd /hydra/TF_data/April2021motifs
#./jasparurl.sh

%%bash
cd /hydra/TF_data/April2021motifs
echo "#!/bin/bash" > jasparurl_lit.sh
echo "while read p; do wget \$p -O \${p:40:6}.meme; done < matrixURLlit.list" >> jasparurl_lit.sh
chmod +x jasparurl_lit.sh
less jasparurl_lit.sh

#%%bash
#cd /hydra/TF_data/April2021motifs
#./jasparurl_lit.sh

%%bash
cd /hydra/TF_data/April2021motifs
#rm all.meme
meme2meme *meme > all.meme

%%bash
cd /hydra/TF_data/April2021motifs
ls -l all.meme

# 4. Final transcription factor table

The TPM values, proportional expression values, motif IDs amd motif evalues were added to the transcription factor table.

import pandas as pd
facs = pd.read_csv("/hydra/FACS/FACS_prop.csv")

TF_table = pd.read_excel("/hydra/TF_data/Transcription_factors_01072021.xlsx")

profiles_filtered1 = pd.read_csv("/hydra/TF_data/jaspar_filtered.csv")
profiles_filtered2 = pd.read_csv("/hydra/TF_data/jaspar_filtered_lit.csv")
profiles_filtered = pd.concat([profiles_filtered1, profiles_filtered2])

TF_table["X1"] = 0
TF_table["X2"] = 0
TF_table["Xins"] = 0
TF_table["X1 percent"] = 0
TF_table["X2 percent"] = 0
TF_table["Xins percent"] = 0
TF_table["Jaspar MatrixID"] = "0"
TF_table["Jaspar Evalue"] = 0
for i in range(len(TF_table)):
    TF_table["X1"].iloc[i] = facs[facs["gene"] == TF_table["Neiro"].iloc[i]]["X1"].iloc[0]
    TF_table["X2"].iloc[i] = facs[facs["gene"] == TF_table["Neiro"].iloc[i]]["X2"].iloc[0]
    TF_table["Xins"].iloc[i] = facs[facs["gene"] == TF_table["Neiro"].iloc[i]]["Xins"].iloc[0]
    TF_table["X1 percent"].iloc[i] = facs[facs["gene"] == TF_table["Neiro"].iloc[i]]["X1.prop"].iloc[0]
    TF_table["X2 percent"].iloc[i] = facs[facs["gene"] == TF_table["Neiro"].iloc[i]]["X2.prop"].iloc[0]
    TF_table["Xins percent"].iloc[i] = facs[facs["gene"] == TF_table["Neiro"].iloc[i]]["Xins.prop"].iloc[0]
    matrixidtable = profiles_filtered[profiles_filtered["GeneID"] == TF_table["Neiro"].iloc[i]]
    if len(matrixidtable) > 0:
        TF_table["Jaspar MatrixID"].iloc[i] = matrixidtable["MatrixID"].iloc[0]
        TF_table["Jaspar Evalue"].iloc[i] = matrixidtable["Evalue"].iloc[0]

TF_table.to_excel("/hydra/TF_data/Transcription_factors_01072021_values.xlsx")
TF_table.to_csv("/hydra/TF_data/Transcription_factors_01072021_values.csv")

## 4.2 Isoforms of transcription factors

import pandas as pd
TF_table = pd.read_csv("/hydra/TF_data/Transcription_factors_01072021_values.csv").iloc[:,1:]

len(TF_table)

len(TF_table[~TF_table["Reference"].isna()])

TF_table[~TF_table["Reference"].isna()]

counts = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/counts_annotated.csv")

counts[counts.iloc[:,0].isin(TF_table["Neiro"])].sort_values(by="Counts", ascending=False).iloc[120:150,:]

TF_table[TF_table["Neiro"] == "MSTRG.442"]

%%bash
cd /hydra/sexual_genome_annotation_files/ncrna_Neiro/
grep "MSTRG.442" stringtie_merged.gtf

# FINNISHED
