==> notebooks/4-2-ATAC-PeakCoverage.ipynb <==
# 4-2-ATAC-PeakCoverage
Jakke Neiro$^1$
1. Aboobaker laboratory, Department of Zoology, University of Oxford

## Contents of notebook
* 1. Introduction
* 2. Coverage 
    * 3.1 X1
    * 3.2 X2
    * 3.3 Xins
    * 3.4 Blastemal ATAC-seq
* 3. Peak calling
    * 3.1 X1
        * 3.1.1 MACS2
        * 3.1.2 Genrich
    * 3.2 X2
        * 3.2.1 MACS2
        * 3.2.2 Genrich
    * 3.3 Xins
        * 3.3.1 MACS2
        * 3.3.2 Genrich
    * 3.4 Blastemal ATAC-seq
        * 3.4.1 MACS2
        * 3.4.2 Genrich

## Files
* Input: x1lysis.marked.bam, x1wcs.marked.bam, x2lysis.marked.bam, x2wcs.marked.bam, xinsATAC.marked.bam, xins2.marked.bam, SRR8441376.marked.bam, SRR8441377.marked.bam, SRR8441378.marked.bam, SRR8441379.marked.bam
* Output: x1coverage_mean_april2021.bw, x2coverage_mean_april2021.bw, xinscoverage_mean_april2021.bw, mansi6h_april2021.bw, mansi0h_april2021.bw, x1atacapril21_summits.bed, x2atacapril21_summits.bed, xinsatacapril21_summits.bed, x1Genrich_april2021.narrowPeak, x2Genrich_april2021.narrowPeak, xinsGenrich_april2021.narrowPeak, 6hapril2021_summits.bed, 0hapril2021_summits.bed, 6hGenrich_april2021.narrowPeak, 0hGenrich_april2021.narrowPeak

# 1. Introduction

Coverage files were generated for ATAC-seq data and peaks were called.

# 2. Coverage

Coverage files were generated for X1, X2, Xins and blastemal ATAC-seq samples.

## 2.1 X1

The coverage for X1 samples was generated: 

%%bash
cd /hydra/atac_sexual_final/x1/
echo "#!/bin/bash" > bamCoverage_mean_April2021.sh
echo "bamCoverage -b x1lysis.marked.bam -o x1lysiscoverage_april2021.bw -of bigwig --normalizeUsing RPKM" >> bamCoverage_mean_April2021.sh
echo "bamCoverage -b x1wcs.marked.bam -o x1wcscoverage_april2021.bw -of bigwig --normalizeUsing RPKM" >> bamCoverage_mean_April2021.sh
echo "bigwigCompare -b1 x1lysiscoverage_april2021.bw -b2 x1wcscoverage_april2021.bw -o x1coverage_mean_april2021.bw -of bigwig --operation mean" >> bamCoverage_mean_April2021.sh
chmod +x bamCoverage_mean_April2021.sh
less bamCoverage_mean_April2021.sh

%%bash
cd /drives/raid/AboobakerLab/jakke/atac_sexual_final/x1/
ls *.marked.bam

#%%bash
#cd /hydra/atac_sexual/x1/
#nohup ./bamCoverage_mean_April2021.sh > nohup_cov.out 2>&1  </dev/null &

%%bash
cd /hydra/atac_sexual_final/x1/
ls x1coverage_mean_april2021.bw

## 2.2 X2

The coverage for X2 samples was generated:

%%bash
cd /hydra/atac_sexual_final/x2/
echo "#!/bin/bash" > bamCoverage_mean_April2021.sh
echo "bamCoverage -b x2lysis.marked.bam -o x2lysiscoverage_april2021.bw -of bigwig --normalizeUsing RPKM" >> bamCoverage_mean_April2021.sh
echo "bamCoverage -b x2wcs.marked.bam -o x2wcscoverage_april2021.bw -of bigwig --normalizeUsing RPKM" >> bamCoverage_mean_April2021.sh
echo "bigwigCompare -b1 x2lysiscoverage_april2021.bw -b2 x2wcscoverage_april2021.bw -o x2coverage_mean_april2021.bw -of bigwig --operation mean" >> bamCoverage_mean_April2021.sh
chmod +x bamCoverage_mean_April2021.sh
less bamCoverage_mean_April2021.sh

#%%bash
#cd /hydra/atac_sexual/x2/
#nohup ./bamCoverage_mean_April2021.sh > nohup_cov.out 2>&1  </dev/null &

%%bash
cd /hydra/atac_sexual_final/x2/
ls x2coverage_mean_april2021.bw

## 2.3 Xins

The coverage for Xins samples was generated: 

%%bash
cd /hydra/atac_sexual_final/xins/
echo "#!/bin/bash" > bamCoverage_mean_April2021.sh
echo "bamCoverage -b xins2.marked.bam -o xins2scoverage_april2021.bw -of bigwig --normalizeUsing RPKM" >> bamCoverage_mean_April2021.sh
echo "bamCoverage -b xinsATAC.marked.bam -o xins1overage_april2021.bw -of bigwig --normalizeUsing RPKM" >> bamCoverage_mean_April2021.sh
echo "bigwigCompare -b1 xins2scoverage_april2021.bw -b2 xins1overage_april2021.bw -o xinscoverage_mean_april2021.bw -of bigwig --operation mean" >> bamCoverage_mean_April2021.sh
chmod +x bamCoverage_mean_April2021.sh
less bamCoverage_mean_April2021.sh

#%%bash
#cd /hydra/atac_sexual_final/xins/
#nohup ./bamCoverage_mean_April2021.sh > nohup_cov.out 2>&1  </dev/null &

%%bash
cd /hydra/atac_sexual_final/xins/
ls xinscoverage_mean_april2021.bw

## 2.4 Blastemal ATAC-seq

The coverage for blastemal ATAC-seq samples was generated: 

%%bash
cd /hydra/atac_sexual_final/mansi_atac/
echo "#!/bin/bash" > bamCoverage_mean_April2021.sh
echo "bamCoverage -b SRR8441376.marked.bam -o SRR8441376_april2021.bw -of bigwig --normalizeUsing RPKM" >> bamCoverage_mean_April2021.sh
echo "bamCoverage -b SRR8441377.marked.bam -o SRR8441377_april2021.bw -of bigwig --normalizeUsing RPKM" >> bamCoverage_mean_April2021.sh
echo "bamCoverage -b SRR8441378.marked.bam -o SRR8441378_april2021.bw -of bigwig --normalizeUsing RPKM" >> bamCoverage_mean_April2021.sh
echo "bamCoverage -b SRR8441379.marked.bam -o SRR8441379_april2021.bw -of bigwig --normalizeUsing RPKM" >> bamCoverage_mean_April2021.sh
chmod +x bamCoverage_mean_April2021.sh
less bamCoverage_mean_April2021.sh

#%%bash
#cd /hydra/atac_sexual_final/mansi_atac/
#nohup ./bamCoverage_mean_April2021.sh > nohup_cov.out 2>&1  </dev/null &

%%bash
cd /hydra/atac_sexual_final/mansi_atac/
wiggletools mean SRR8441376_april2021.bw SRR8441377_april2021.bw > mansi6h_2021.wig

%%bash
cd /hydra/atac_sexual_final/mansi_atac/
wiggletools mean SRR8441378_april2021.bw SRR8441379_april2021.bw > mansi0h_2021.wig

%%bash
cd /hydra/atac_sexual_final/mansi_atac/
wigToBigWig mansi6h_2021.wig /hydra/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome mansi6h_april2021.bw

%%bash
cd /hydra/atac_sexual_final/mansi_atac/
wigToBigWig mansi0h_2021.wig /hydra/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome mansi0h_april2021.bw

%%bash
cd /hydra/atac_sexual_final/mansi_atac/
ls mansi6h_april2021.bw
ls mansi0h_april2021.bw

# 3. Peak calling 

## 3.1 X1

### 3.1.1 MACS2

%%bash
cd /hydra/atac_sexual_final/x1/
echo "#!/bin/bash" > macs2peak_april2021.sh
echo "macs2 callpeak -t x1lysis.marked.bam x1wcs.marked.bam --nomodel --extsize 200 --shift -100 -n x1atacapril21" >> macs2peak_april2021.sh
chmod +x macs2peak_april2021.sh
less macs2peak_april2021.sh

#%%bash
#cd /hydra/atac_sexual_final/x1/
#nohup ./macs2peak_april2021.sh > nohup_cov.out 2>&1  </dev/null &

%%bash
cd /hydra/atac_sexual_final/x1/
wc -l x1atacapril21_summits.bed

### 3.1.2 Genrich

%%bash
cd /hydra/atac_sexual_final/x1/
samtools sort -o x1lysis.marked.sorted.bam -n x1lysis.marked.bam
samtools sort -o x1wcs.marked.sorted.bam -n x1wcs.marked.bam 

%%bash
cd /hydra/atac_sexual_final/x1/
echo "#!/bin/bash" > genrichpeak_april2021.sh
echo "/hydra/software/Genrich/Genrich -t x1lysis.marked.sorted.bam,x1wcs.marked.sorted.bam -o x1Genrich_april2021.narrowPeak -f x1Genrich_april2021.bed -y -j -v" >> genrichpeak_april2021.sh
chmod +x genrichpeak_april2021.sh
less genrichpeak_april2021.sh

#%%bash
#cd /hydra/atac_sexual_final/x1/
#nohup ./genrichpeak_april2021.sh > nohup_cov.out 2>&1  </dev/null &

%%bash
cd /hydra/atac_sexual_final/x1/
wc -l x1Genrich_april2021.narrowPeak

## 3.2 X2

### 3.2.1 MACS2

%%bash
cd /hydra/atac_sexual_final/x2/
echo "#!/bin/bash" > macs2peak_april2021.sh
echo "macs2 callpeak -t x2lysis.marked.bam x2wcs.marked.bam --nomodel --extsize 200 --shift -100 -n x2atacapril21" >> macs2peak_april2021.sh
chmod +x macs2peak_april2021.sh
less macs2peak_april2021.sh

#%%bash
#cd /hydra/atac_sexual_final/x2/
#nohup ./macs2peak_april2021.sh > nohup_cov.out 2>&1  </dev/null &

%%bash
cd /hydra/atac_sexual_final/x2/
wc -l x2atacapril21_summits.bed

### 3.2.2 Genrich

%%bash
cd /hydra/atac_sexual_final/x2/
echo "#!/bin/bash" > genrichpeak_april2021.sh
echo "samtools sort -o x2lysis.marked.sorted.bam -n x2lysis.marked.bam" >> genrichpeak_april2021.sh
echo "samtools sort -o x2wcs.marked.sorted.bam -n x2wcs.marked.bam" >> genrichpeak_april2021.sh
echo "/hydra/software/Genrich/Genrich -t x2lysis.marked.sorted.bam,x2wcs.marked.sorted.bam -o x2Genrich_april2021.narrowPeak -f x2Genrich_april2021.bed -y -j -v" >> genrichpeak_april2021.sh
chmod +x genrichpeak_april2021.sh
less genrichpeak_april2021.sh

#%%bash
#cd /hydra/atac_sexual_final/x2/
#nohup ./genrichpeak_april2021.sh > nohup_cov.out 2>&1  </dev/null &

%%bash
cd /hydra/atac_sexual_final/x2/
wc -l x2Genrich_april2021.narrowPeak

## 3.3 Xins

### 3.3.1 MACS2

%%bash
cd /hydra/atac_sexual_final/xins/
echo "#!/bin/bash" > macs2peak_april2021.sh
echo "macs2 callpeak -t xins2.marked.bam xinsATAC.marked.bam --nomodel --extsize 200 --shift -100 -n xinsatacapril21" >> macs2peak_april2021.sh
chmod +x macs2peak_april2021.sh
less macs2peak_april2021.sh

#%%bash
#cd /hydra/atac_sexual_final/xins/
#nohup ./macs2peak_april2021.sh > nohup_cov.out 2>&1  </dev/null &

%%bash
cd /hydra/atac_sexual_final/xins/
wc -l xinsatacapril21_summits.bed

### 3.3.2 Genrich

%%bash
cd /hydra/atac_sexual_final/xins/
echo "#!/bin/bash" > genrichpeak_april2021.sh
echo "samtools sort -o xins2.marked.sorted.bam -n xins2.marked.bam" >> genrichpeak_april2021.sh
echo "samtools sort -o xinsATAC.marked.sorted.bam -n xinsATAC.marked.bam" >> genrichpeak_april2021.sh
echo "/hydra/software/Genrich/Genrich -t xins2.marked.sorted.bam,xinsATAC.marked.sorted.bam -o xinsGenrich_april2021.narrowPeak -f xinsGenrich_april2021.bed -y -j -v" >> genrichpeak_april2021.sh
chmod +x genrichpeak_april2021.sh
less genrichpeak_april2021.sh

#%%bash
#cd /hydra/atac_sexual_final/xins/
#nohup ./genrichpeak_april2021.sh > nohup_cov.out 2>&1  </dev/null &

%%bash
cd /hydra/atac_sexual_final/xins/
wc -l xinsGenrich_april2021.narrowPeak

## 3.4 Blastemal ATAC-seq

### 3.4.1 MACS2

%%bash
cd /hydra/atac_sexual_final/mansi_atac/
echo "#!/bin/bash" > macs2peak_april2021.sh
echo "macs2 callpeak -t SRR8441376.marked.bam SRR8441377.marked.bam --nomodel --extsize 200 --shift -100 -n 6hapril2021" >> macs2peak_april2021.sh
echo "macs2 callpeak -t SRR8441378.marked.bam SRR8441379.marked.bam --nomodel --extsize 200 --shift -100 -n 0hapril2021" >> macs2peak_april2021.sh
chmod +x macs2peak_april2021.sh
less macs2peak_april2021.sh

#%%bash
#cd /hydra/atac_sexual_final/mansi_atac/
#nohup ./macs2peak_april2021.sh > nohup_cov.out 2>&1  </dev/null &

%%bash
cd /hydra/atac_sexual_final/mansi_atac/
wc -l 6hapril2021_summits.bed
wc -l 0hapril2021_summits.bed

### 3.4.2 Genrich

%%bash
cd /hydra/atac_sexual_final/mansi_atac/
echo "#!/bin/bash" > genrichpeak_april2021.sh
echo "samtools sort -o SRR8441376.marked.sorted.bam -n SRR8441376.marked.bam" >> genrichpeak_april2021.sh
echo "samtools sort -o SRR8441377.marked.sorted.bam -n SRR8441377.marked.bam" >> genrichpeak_april2021.sh
echo "samtools sort -o SRR8441378.marked.sorted.bam -n SRR8441378.marked.bam" >> genrichpeak_april2021.sh
echo "samtools sort -o SRR8441379.marked.sorted.bam -n SRR8441379.marked.bam" >> genrichpeak_april2021.sh
echo "/hydra/software/Genrich/Genrich -t SRR8441376.marked.sorted.bam,SRR8441377.marked.sorted.bam -o 6hGenrich_april2021.narrowPeak -f 6hGenrich_april2021.bed -y -j -v" >> genrichpeak_april2021.sh
echo "/hydra/software/Genrich/Genrich -t SRR8441378.marked.sorted.bam,SRR8441379.marked.sorted.bam -o 0hGenrich_april2021.narrowPeak -f 0hGenrich_april2021.bed -y -j -v" >> genrichpeak_april2021.sh
chmod +x genrichpeak_april2021.sh
less genrichpeak_april2021.sh

#%%bash
#cd /hydra/atac_sexual_final/mansi_atac/
#nohup ./genrichpeak_april2021.sh > nohup_cov.out 2>&1  </dev/null &

%%bash
cd /hydra/atac_sexual_final/mansi_atac/
wc -l 6hGenrich_april2021.narrowPeak
wc -l 0hGenrich_april2021.narrowPeak

# FINNISHED
