==> notebooks/4-3-ATAC-footprinting.ipynb <==
# 4-3-ATAC-Footprinting
Jakke Neiro$^1$
1. Aboobaker laboratory, Department of Zoology, University of Oxford

## Contents of notebook
* 1. Introduction
* 2. Footprinting TOBIAS
    * 2.1 ATACCorrect
        * 2.1.1 X1
        * 2.1.2 X2
        * 2.1.3 Xins
     * 2.2 ScoreBigWig
        * 2.2.1 X1
        * 2.2.2 X2
        * 2.2.3 Xins
        * 2.2.4 
    * 2.3 BindDetect
        * 2.3.1 X1
        * 2.3.2 X1 vs X2 vs Xins

## Files
* Input:
* Output:

# 1. Introduction

# 2. Footprinting TOBIAS

## 2.1 ATACCorrect

Let's first check the files:

%%bash
cd /hydra/enhancer_regions
ls -l chip_summits_july21.bed
head -2 chip_summits_july21.bed

import pandas as pd
chip_summits = pd.read_csv("/hydra/enhancer_regions/chip_summits_july21.bed", sep="\t", header=None)
chip_summits["Start"] = (chip_summits.iloc[:,1]+chip_summits.iloc[:,2])//2
chip_summits["Stop"] = chip_summits["Start"] + 1

chip_summits.iloc[:,[0,5,6]].to_csv("/hydra/enhancer_regions/chip_peaks_july21.bed", sep="\t", index=False, header=None)

%%bash
cd /hydra/enhancer_regions
bedtools slop -i /hydra/enhancer_regions/chip_peaks_july21.bed -g /hydra/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome -b 1000 > /hydra/enhancer_regions/chip_summits_regions.bed

%%bash
cd /hydra/enhancer_regions
bedtools slop -i /hydra/enhancer_regions/random_peaks.bed -g /hydra/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome -b 1000 > /hydra/enhancer_regions/random_regions.bed

%%bash
cd /hydra/enhancer_regions
wc -l chip_summits_regions.bed
head -2 chip_summits_regions.bed
head -2 random_regions.bed

### 2.1.1 X1 

%%bash
cd /hydra/atac_sexual_final
echo "#!/bin/bash" > x1_ataccorrect_april21regions.sh
echo "cd /hydra/atac_sexual_final/x1" >> x1_ataccorrect_april21regions.sh
echo "TOBIAS ATACorrect --bam x1lysis.marked.bam --genome /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa --peaks /hydra/enhancer_regions/chip_summits_regions.bed --outdir /hydra/atac_sexual_final/x1_lysis_atacorrect_regions --cores 2" >> x1_ataccorrect_april21regions.sh
echo "TOBIAS ATACorrect --bam x1wcs.marked.bam --genome /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa --peaks /hydra/enhancer_regions/chip_summits_regions.bed --outdir /hydra/atac_sexual_final/x1_wcs_atacorrect_regions --cores 2" >> x1_ataccorrect_april21regions.sh
chmod +x x1_ataccorrect_april21regions.sh
less x1_ataccorrect_april21regions.sh

#%%bash
#cd /hydra/atac_sexual_final
#nohup ./x1_ataccorrect_april21regions.sh &

### 2.1.2 X2

%%bash
cd /hydra/atac_sexual_final
echo "#!/bin/bash" > x2_ataccorrect_april21regions.sh
echo "cd /hydra/atac_sexual_final/x2" >> x2_ataccorrect_april21regions.sh
echo "TOBIAS ATACorrect --bam x2lysis.marked.bam --genome /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa --peaks /hydra/enhancer_regions/chip_summits_regions.bed --outdir /hydra/atac_sexual_final/x2_lysis_atacorrect_regions --cores 2" >> x2_ataccorrect_april21regions.sh
echo "TOBIAS ATACorrect --bam x2wcs.marked.bam --genome /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa --peaks /hydra/enhancer_regions/chip_summits_regions.bed --outdir /hydra/atac_sexual_final/x2_wcs_atacorrect_regions --cores 2" >> x2_ataccorrect_april21regions.sh
chmod +x x2_ataccorrect_april21regions.sh
less x2_ataccorrect_april21regions.sh

#%%bash
#cd /hydra/atac_sexual_final
#nohup ./x2_ataccorrect_april21regions.sh &

### 2.1.3 Xins

%%bash
cd /hydra/atac_sexual_final
echo "#!/bin/bash" > xins_ataccorrect_april21regions.sh
echo "cd /hydra/atac_sexual_final/xins" >> xins_ataccorrect_april21regions.sh
echo "TOBIAS ATACorrect --bam xins2.marked.bam --genome /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa --peaks /hydra/enhancer_regions/chip_summits_regions.bed --outdir /hydra/atac_sexual_final/xins_2_atacorrect_regions --cores 2" >> xins_ataccorrect_april21regions.sh
echo "TOBIAS ATACorrect --bam xinsATAC.marked.bam --genome /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa --peaks /hydra/enhancer_regions/chip_summits_regions.bed --outdir /hydra/atac_sexual_final/xins_1_atacorrect_regions --cores 2" >> xins_ataccorrect_april21regions.sh
chmod +x xins_ataccorrect_april21regions.sh
less xins_ataccorrect_april21regions.sh

#%%bash
#cd /hydra/atac_sexual_final
#nohup ./xins_ataccorrect_april21regions.sh &

### 2.1.4 Random peaks

%%bash
cd /hydra/atac_sexual_final
echo "#!/bin/bash" > randomx1_ataccorrect_april21regions.sh
echo "cd /hydra/atac_sexual_final/x1" >> randomx1_ataccorrect_april21regions.sh
echo "TOBIAS ATACorrect --bam x1lysis.marked.bam --genome /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa --peaks /hydra/enhancer_regions/random_regions.bed --outdir /hydra/atac_sexual_final/randomx1_lysis_atacorrect_regions --cores 8" >> randomx1_ataccorrect_april21regions.sh
echo "TOBIAS ATACorrect --bam x1wcs.marked.bam --genome /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa --peaks /hydra/enhancer_regions/random_regions.bed --outdir /hydra/atac_sexual_final/randomx1_wcs_atacorrect_regions --cores 8" >> randomx1_ataccorrect_april21regions.sh
chmod +x randomx1_ataccorrect_april21regions.sh
less randomx1_ataccorrect_april21regions.sh

#%%bash
#cd /hydra/atac_sexual_final
#nohup ./randomx1_ataccorrect_april21regions.sh &

## 2.2 ScoreBigWig

### 2.2.1 X1

%%bash
cd /hydra/atac_sexual_final
echo "#!/bin/bash" > x1_scorebigwig_april21regions.sh
echo "cd /hydra/atac_sexual_final/footprinting" >> x1_scorebigwig_april21regions.sh
echo "TOBIAS FootprintScores --signal /hydra/atac_sexual_final/x1_lysis_atacorrect_regions/x1lysis.marked_corrected.bw --regions /hydra/enhancer_regions/chip_summits_regions.bed --output x1lysismarked_footprints_regions.bw --cores 2" >> x1_scorebigwig_april21regions.sh
echo "TOBIAS FootprintScores --signal /hydra/atac_sexual_final/x1_wcs_atacorrect_regions/x1wcs.marked_corrected.bw --regions /hydra/enhancer_regions/chip_summits_regions.bed --output x1wcsmarked_footprints_regions.bw --cores 2" >> x1_scorebigwig_april21regions.sh
chmod +x x1_scorebigwig_april21regions.sh
less x1_scorebigwig_april21regions.sh

#%%bash
#cd /hydra/atac_sexual_final
#nohup ./x1_scorebigwig_april21regions.sh &

%%bash
cd /hydra/atac_sexual_final/footprinting
wiggletools mean x1lysismarked_footprints_regions.bw  x1wcsmarked_footprints_regions.bw > x1marked_footprints_regions.wig

%%bash
cd /hydra/atac_sexual_final/footprinting
wigToBigWig x1marked_footprints_regions.wig /hydra/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome x1marked_footprints_regions.bw

### 2.2.2 X2

%%bash
cd /hydra/atac_sexual_final
echo "#!/bin/bash" > x2_scorebigwig_april21regions.sh
echo "cd /hydra/atac_sexual_final/footprinting" >> x2_scorebigwig_april21regions.sh
echo "TOBIAS FootprintScores --signal /hydra/atac_sexual_final/x2_lysis_atacorrect_regions/x2lysis.marked_corrected.bw --regions /hydra/enhancer_regions/chip_summits_regions.bed --output x2lysismarked_footprints_regions.bw --cores 2" >> x2_scorebigwig_april21regions.sh
echo "TOBIAS FootprintScores --signal /hydra/atac_sexual_final/x2_wcs_atacorrect_regions/x2wcs.marked_corrected.bw --regions /hydra/enhancer_regions/chip_summits_regions.bed --output x2wcsmarked_footprints_regions.bw --cores 2" >> x2_scorebigwig_april21regions.sh
chmod +x x2_scorebigwig_april21regions.sh
less x2_scorebigwig_april21regions.sh

#%%bash
#cd /hydra/atac_sexual_final
#nohup ./x2_scorebigwig_april21regions.sh &

%%bash
cd /hydra/atac_sexual_final/footprinting
wiggletools mean x2lysismarked_footprints_regions.bw  x2wcsmarked_footprints_regions.bw > x2marked_footprints_regions.wig

%%bash
cd /hydra/atac_sexual_final/footprinting
wigToBigWig x2marked_footprints_regions.wig /hydra/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome x2marked_footprints_regions.bw

### 2.2.3 Xins

%%bash
cd /hydra/atac_sexual_final
echo "#!/bin/bash" > xins_scorebigwig_april21regions.sh
echo "cd /hydra/atac_sexual_final/footprinting" >> xins_scorebigwig_april21regions.sh
echo "TOBIAS FootprintScores --signal /hydra/atac_sexual_final/xins_1_atacorrect_regions/xinsATAC.marked_corrected.bw --regions /hydra/enhancer_regions/chip_summits_regions.bed --output xins1marked_footprints_regions.bw --cores 2" >> xins_scorebigwig_april21regions.sh
echo "TOBIAS FootprintScores --signal /hydra/atac_sexual_final/xins_2_atacorrect_regions/xins2.marked_corrected.bw --regions /hydra/enhancer_regions/chip_summits_regions.bed --output xins2marked_footprints_regions.bw --cores 2" >> xins_scorebigwig_april21regions.sh
chmod +x xins_scorebigwig_april21regions.sh
less xins_scorebigwig_april21regions.sh

#%%bash
#cd /hydra/atac_sexual_final
#nohup ./xins_scorebigwig_april21regions.sh &

%%bash
cd /hydra/atac_sexual_final/footprinting
wiggletools mean xins1marked_footprints_regions.bw xins2marked_footprints_regions.bw > xinsmarked_footprints_regions.wig

%%bash
cd /hydra/atac_sexual_final/footprinting
wigToBigWig xinsmarked_footprints_regions.wig /hydra/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome xinsmarked_footprints_regions.bw

### 2.2.4 Random peaks

%%bash
cd /hydra/atac_sexual_final
echo "#!/bin/bash" > randomx1_scorebigwig_april21regions.sh
echo "cd /hydra/atac_sexual_final/footprinting" >> randomx1_scorebigwig_april21regions.sh
echo "TOBIAS FootprintScores --signal /hydra/atac_sexual_final/randomx1_lysis_atacorrect_regions/x1lysis.marked_corrected.bw --regions /hydra/enhancer_regions/random_regions.bed --output randomx1lysismarked_footprints_regions.bw --cores 8" >> randomx1_scorebigwig_april21regions.sh
echo "TOBIAS FootprintScores --signal /hydra/atac_sexual_final/randomx1_wcs_atacorrect_regions/x1wcs.marked_corrected.bw --regions /hydra/enhancer_regions/random_regions.bed --output randomx1wcsmarked_footprints_regions.bw --cores 8" >> randomx1_scorebigwig_april21regions.sh
chmod +x randomx1_scorebigwig_april21regions.sh
less randomx1_scorebigwig_april21regions.sh

#%%bash
#cd /hydra/atac_sexual_final
#nohup ./randomx1_scorebigwig_april21regions.sh &

#%%bash
#cd /hydra/atac_sexual_final/footprinting
#wiggletools mean randomx1lysismarked_footprints_regions.bw randomx1wcsmarked_footprints_regions.bw > randomx1_footprints_regions.wig

#%%bash
#cd /hydra/atac_sexual_final/footprinting
#wigToBigWig randomx1_footprints_regions.wig /hydra/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome randomx1_footprints_regions.bw

## 2.3 BindDetect

%%bash
cd /hydra/TF_data/April2021motifs
ls -l all.meme

### 2.3.1 X1

%%bash
cd /hydra/atac_sexual_final
echo "#!/bin/bash" > x1binddetect_april21regions.sh
echo "cd /hydra/atac_sexual_final/footprinting" >> x1binddetect_april21regions.sh
echo "TOBIAS BINDetect --motifs /hydra/TF_data/April2021motifs/all.meme --signals x1marked_footprints_regions.bw --genome /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa --peaks /hydra/enhancer_regions/chip_summits_regions.bed --outdir BINDetect_X1 --cond_names X1" >> x1binddetect_april21regions.sh
chmod +x x1binddetect_april21regions.sh 
less x1binddetect_april21regions.sh

#%%bash
#cd /hydra/atac_sexual_final
#nohup ./x1binddetect_april21regions.sh &

%%bash
cd /hydra/atac_sexual_final
echo "#!/bin/bash" > randomx1binddetect_april21regions.sh
echo "cd /hydra/atac_sexual_final/footprinting" >> randomx1binddetect_april21regions.sh
echo "TOBIAS BINDetect --motifs /hydra/TF_data/April2021motifs/all.meme --signals randomx1_footprints_regions.bw --genome /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa --peaks /hydra/enhancer_regions/random_regions.bed --outdir BINDetect_random --cond_names X1" >> randomx1binddetect_april21regions.sh
chmod +x randomx1binddetect_april21regions.sh 
less randomx1binddetect_april21regions.sh

#%%bash
#cd /hydra/atac_sexual_final
#nohup ./randomx1binddetect_april21regions.sh &

%%bash
cd /hydra/atac_sexual_final/footprinting/BINDetect_random
ls -l bindetect_figures.pdf

### 2.3.2 X1 vs X2 vs Xins

All X1, X2 and Xins: 

%%bash
cd /hydra/atac_sexual_final
echo "#!/bin/bash" > binddetect_april21.sh
echo "cd /hydra/atac_sexual_final/footprinting" >> binddetect_april21.sh
echo "TOBIAS BINDetect --motifs /hydra/TF_data/April2021motifs/all.meme --signals x1marked_footprints_regions.bw x2marked_footprints_regions.bw xinsmarked_footprints_regions.bw --genome /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa --peaks /hydra/enhancer_regions/chip_summits_regions.bed --outdir BINDetect_output --cond_names X1 X2 Xins --cores 2" >> binddetect_april21.sh
chmod +x binddetect_april21.sh 
less binddetect_april21.sh 

#%%bash
#cd /hydra/atac_sexual_final
#nohup ./binddetect_april21.sh &

# 3. Footprinting result

%%bash
cd /hydra/atac_sexual_final/footprinting/BINDetect_X1
head -2 bindetect_results.txt

* TF_name: name
* total_tfbs: number of binding sites found in input --peaks
* (condition) bound: number of sites predicted bound in the given condition. This is estimated independently per condition based on the distribution of scores, and is therefore very dependent on how well the threshold for bound/unbound was set. If conflict, the change score is the nore correct metric to use.

%%bash
cd /hydra/atac_sexual_final/footprinting/BINDetect_X1/Alx4_MA0853.1/beds
ls

%%bash
cd /hydra/atac_sexual_final/footprinting/BINDetect_X1/Alx4_MA0853.1
head -2 Alx4_MA0853.1_overview.txt

%%bash
cd /hydra/atac_sexual_final/footprinting/BINDetect_X1/Alx4_MA0853.1/beds
head -2 Alx4_MA0853.1_X1_bound.bed

import pandas as pd
X1bind = pd.read_csv("/hydra/atac_sexual_final/footprinting/BINDetect_X1/bindetect_results.txt", sep="\t")

X1bind 

tf_prop = pd.read_csv("/hydra/TF_data/Transcription_factors_01072021_values.csv").iloc[:,1:]
X1bind["Name"] = "-"
for i in range(len(X1bind)):
    matrix_id = X1bind.iloc[i, 0].split("_")[1]
    gene_id_table = tf_prop[tf_prop["Jaspar MatrixID"] == matrix_id]
    if len(gene_id_table) > 0:
        name_str = gene_id_table.iloc[0,2]
        for j in range(1,len(gene_id_table)):
            name_str = name_str + ", " + gene_id_table.iloc[j,2]
    X1bind["Name"].iloc[i] = name_str

X1bind.sort_values(by="X1_bound", ascending=False).to_excel("/hydra/atac_sexual_final/footprinting/BINDetect_X1/bindetect_results_names.xlsx", index=False) 

tf_prop = pd.read_csv("/hydra/TF_data/Transcription_factors_01072021_values.csv").iloc[:,1:]

X1bind

tf_prop = pd.read_csv("/hydra/TF_data/Transcription_factors_01072021_values.csv").iloc[:,1:]

%%bash
cd /hydra/atac_sexual_final/footprinting/BINDetect_X1
ls -d *M* > dict.list

dicts = pd.read_csv("/hydra/atac_sexual_final/footprinting/BINDetect_X1/dict.list", header=None)

motif_dict = dicts.iloc[0, 0]
table = pd.read_csv("/hydra/atac_sexual_final/footprinting/BINDetect_X1/"+motif_dict+"/"+motif_dict+"_overview.txt", sep="\t")
for i in range(1, len(dicts)):
    motif_dict = dicts.iloc[i, 0]
    table1 = pd.read_csv("/hydra/atac_sexual_final/footprinting/BINDetect_X1/"+motif_dict+"/"+motif_dict+"_overview.txt", sep="\t")
    table = pd.concat([table, table1])

X1_table = table

len(X1_table)

X1_table["GeneID"] = "a"
X1_table["TF name"] = "a"
for i in range(0, len(X1_table)):
    matrix_id = X1_table.iloc[i, 3].split("_")[1]
    gene_id_table = tf_prop[tf_prop["Jaspar MatrixID"] == matrix_id]
    if len(gene_id_table) == 0:
        continue
    gene_id_str = gene_id_table.iloc[0,1]
    name_id_str = gene_id_table.iloc[0,2]
    for j in range(1,len(gene_id_table)):
        gene_id_str = gene_id_str + ":" + gene_id_table.iloc[j,1]
        name_id_str = name_id_str + ":" + gene_id_table.iloc[j,2]
    X1_table["GeneID"].iloc[i] = gene_id_str
    X1_table["TF name"].iloc[i] = name_id_str
    if i%10000 == 1:
        print(i)
X1_table.to_csv("/hydra/atac_sexual_final/footprinting/BINDetect_X1/X1_table.csv", index=False)

import pandas as pd
X1_table = pd.read_csv("/hydra/atac_sexual_final/footprinting/BINDetect_X1/X1_table.csv")

X1_table.head()

## X1 vs X2 vs Xins

%%bash
cd /hydra/atac_sexual_final/footprinting/BINDetect_output
ls -d *M* > dict.list

import pandas as pd
dicts = pd.read_csv("/hydra/atac_sexual_final/footprinting/BINDetect_output/dict.list", header=None)

motif_dict = dicts.iloc[0, 0]
table = pd.read_csv("/hydra/atac_sexual_final/footprinting/BINDetect_output/"+motif_dict+"/"+motif_dict+"_overview.txt", sep="\t")
for i in range(1, len(dicts)):
    motif_dict = dicts.iloc[i, 0]
    table1 = pd.read_csv("/hydra/atac_sexual_final/footprinting/BINDetect_output/"+motif_dict+"/"+motif_dict+"_overview.txt", sep="\t")
    table = pd.concat([table, table1])

X1X2Xins_table = table

X1X2Xins_table.head()

X1X2Xins_table["GeneID"] = "a"
X1X2Xins_table["TF name"] = "a"
for i in range(0, len(X1X2Xins_table)):
    matrix_id = X1X2Xins_table.iloc[i, 3].split("_")[1]
    gene_id_table = tf_prop[tf_prop["Jaspar MatrixID"] == matrix_id]
    if len(gene_id_table) == 0:
        continue
    gene_id_str = gene_id_table.iloc[0,1]
    name_id_str = gene_id_table.iloc[0,2]
    for j in range(1,len(gene_id_table)):
        gene_id_str = gene_id_str + ":" + gene_id_table.iloc[j,1]
        name_id_str = name_id_str + ":" + gene_id_table.iloc[j,2]
    X1X2Xins_table["GeneID"].iloc[i] = gene_id_str
    X1X2Xins_table["TF name"].iloc[i] = name_id_str
    if i%10000 == 1:
        print(i)
X1X2Xins_table.to_csv("/hydra/atac_sexual_final/footprinting/BINDetect_output/X1X2Xins_table.csv", index=False)

# foxG and Wnt

%%bash
cd /hydra/atac_sexual_final/footprinting
echo -e "dd_Smes_g4_197\t264500\t265500" > wnt_enhancer.bed
echo -e "dd_Smes_g4_197\t266000\t267000" >> wnt_enhancer.bed
less wnt_enhancer.bed

SMESG000024191

%%bash
cd /hydra/TF_data/April2021motifs
wget http://jaspar.genereg.net/api/v1/matrix/MA0458.1/?format=meme -O MA0458.meme

%%bash
cd /hydra/atac_sexual_final
echo "#!/bin/bash" > x1binddetect_wnts.sh
echo "cd /hydra/atac_sexual_final/footprinting" >> x1binddetect_wnts.sh
echo "TOBIAS BINDetect --motifs /hydra/TF_data/April2021motifs/MA0458.meme --signals x1marked_footprints_regions.bw --genome /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa --peaks /hydra/enhancer_regions/chip_summits_regions.bed --outdir BINDetect_wnt --cond_names X1" >> x1binddetect_wnts.sh
chmod +x x1binddetect_wnts.sh
less x1binddetect_wnts.sh

%%bash
ls /hydra/atac_sexual_final/footprinting/BINDetect_wnt/slp1_MA0458.1/beds/

import pandas as pd
X1_table = pd.read_csv("/hydra/atac_sexual_final/footprinting/BINDetect_wnt/slp1_MA0458.1/beds/slp1_MA0458.1_X1_unbound.bed", sep="\t", header=None)
X1_table.iloc[:,0:3].to_csv("/hydra/atac_sexual_final/footprinting/BINDetect_wnt/slp1_MA0458.1/beds/slp1.bed", index=False, header=None, sep="\t")

ndl-3, ptk7, and trig-1
