==> notebooks/5-1-MBD34RNAi-H3K4me1ChIP.ipynb <==
# 5-1-MBD34RNAi-H3K4me1ChIP
Jakke Neiro$^1$

1. Aboobaker laboratory, Department of Zoology, University of Oxford

## Contents of notebook

* 1. Introduction
* 2. Data quality
    * 2.1 Data download
    * 2.2 FastQC
        * 2.2.1 SRR4089759 (h3k4me1_lpt02)
        * 2.2.2 SRR4089760 (h3k4me1_lpt03)
        * 2.2.3 SRR4089771 (input_lpt01)
        * 2.2.4 SRR4089772 (input_lpt02)
        * 2.2.5 SRR4089773 (input_lpt03)
        * 2.2.6 SRR4089846 (h3k4me1_lpt01)
* 3. Trimming
    * 3.1 Trimmomatic
    * 3.2 FastQC after trimming
        * 3.2.1 SRR4089759 (h3k4me1_lpt02)
        * 3.2.2 SRR4089760 (h3k4me1_lpt03)
        * 3.2.3 SRR4089771 (input_lpt01)
        * 3.2.4 SRR4089772 (input_lpt02)
        * 3.2.5 SRR4089773 (input_lpt03)
        * 3.2.6 SRR4089846 (h3k4me1_lpt01)
* 4. Alignment
* 5. Filtering
    * 5.1 Sam to Bam
    * 5.2 Unique reads and sorting
    * 5.3 Removing duplicates
    * 5.4 Read retainment
        * 5.4.1 SRR4089759 (h3k4me1_lpt02)
        * 5.4.2 SRR4089760 (h3k4me1_lpt03)
        * 5.4.3 SRR4089771 (input_lpt01)
        * 5.4.4 SRR4089772 (input_lpt02)
        * 5.4.5 SRR4089773 (input_lpt03)
        * 5.4.6 SRR4089846 (h3k4me1_lpt01)
* 6. Coverage tracks
    * 6.1 ChIP-seq H3K4me1 track
    * 6.2 ChIP-seq H3K4me1 WT vs. lpt(RNAi) track

## Files

* Input: Fastq files: SRR4089846, SRR4089759, SRR4089760, SRR4089771, SRR4089772, SRR4089773
* Output: h3k4me1RNAilog2.bw

# 1. Introduction

The H3K4me1 ChIP-seq data upon MBD3/4 RNAi was inspected, trimmed, and aligned to the planarian genome. Coverage files were generated. 

* Monomethylation of histone H3K4 (H3K4me1) is relatively more enriched at metazoan enhancer regions compared to trimethylated histone H3K4 (H3K4me3), which is enriched at transcription start sites in all eukaryotes.

# 2. Data quality

## 2.1 Data download

%%bash
cd /hydra/chip_RNAi
less sample.txt

%%bash
cd /hydra/chip_RNAi
less data_download.sh

## 2.2 FastQC

%%bash
cd /hydra/chip_RNAi
less fastqc_paired.sh
#Script for FastQC of all samples

#%%bash
#cd /hydra/chip_RNAi
#nohup ./fastqc_paired.sh &
#Running script

### 2.2.1 SRR4089759 (h3k4me1_lpt02)

%%bash
cd /hydra/chip_RNAi/SRR4089759_1_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/SRR4089759_1_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

%%bash
cd /hydra/chip_RNAi/SRR4089759_2_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/SRR4089759_2_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

There are overrepresented sequences to be removed

### 2.2.2 SRR4089760 (h3k4me1_lpt03)

%%bash
cd /hydra/chip_RNAi/SRR4089760_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/SRR4089760_fastqc
less summary.txt

### 2.2.3 SRR4089771 (input_lpt01)

%%bash
cd /hydra/chip_RNAi/SRR4089771_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/SRR4089771_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.4 SRR4089772 input_lpt02

%%bash
cd /hydra/chip_RNAi/SRR4089772_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/SRR4089772_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.5 SRR4089773 (input_lpt03)

%%bash
cd /hydra/chip_RNAi/SRR4089773_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/SRR4089773_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.6 SRR4089846 (h3k4me1_lpt01)

%%bash
cd /hydra/chip_RNAi/SRR4089846_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/SRR4089846_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

# 3. Trimming

%%bash
cd /hydra/software/Trimmomatic-0.39/adapters
less TruSeq3-SE.fa

%%bash
cd /hydra/chip_RNAi/
less TruSeq9.fa
less TruSeq5.fa
less TruSeq7.fa
less TruSeq3.fa
less TruSeq11.fa
less TruSeq1.fa

## 3.1 Trimmomatic

%%bash
cd /hydra/chip_RNAi/
less trimmomatic_paired.sh
#Script for trimming with trimmomatic

#%%bash
#cd /hydra/chip_RNAi/
#nohup ./trimmomatic_paired.sh &
#Running script

## 3.2 FastQC after trimming

#%%bash
#cd /hydra/chip_RNAi/
#nohup fastqc SRR4089759.trimmed.fastq.gz SRR4089760.trimmed.fastq.gz SRR4089771.trimmed.fastq.gz SRR4089772.trimmed.fastq.gz SRR4089773.trimmed.fastq.gz SRR4089846.trimmed.fastq.gz -o ~/chip/ &

### 3.2.1 SRR4089759 (h3k4me1_lpt02)

%%bash
cd /hydra/chip_RNAi/SRR4089759.trimmed_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/SRR4089759.trimmed_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 3.2.2 SRR4089760 (h3k4me1_lpt03)

%%bash
cd /hydra/chip_RNAi/SRR4089760.trimmed_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/SRR4089760.trimmed_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 3.2.3 SRR4089771 (input_lpt01)

%%bash
cd /hydra/chip_RNAi/SRR4089771.trimmed_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/SRR4089771.trimmed_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 3.2.4 SRR4089772 input_lpt02

%%bash
cd /hydra/chip_RNAi/SRR4089772.trimmed_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/SRR4089772.trimmed_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 3.2.5 SRR4089773 (input_lpt03)

%%bash
cd /hydra/chip_RNAi/SRR4089773.trimmed_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/SRR4089773.trimmed_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 3.2.6 SRR4089846 (h3k4me1_lpt01)

%%bash
cd /hydra/chip_RNAi/SRR4089846.trimmed_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/SRR4089846.trimmed_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

# 4. Alignment

%%bash
cd /hydra/chip_RNAi/
less index.sh

#%%bash
#cd ~/../../../../data/ngs/smed/facs_rnaseq_datasets/sexual/chip_RNAi/
#nohup ./index.sh &

%%bash
cd /hydra/chip_RNAi
less align_paired.sh

#%%bash
#cd /hydra/chip_RNAi
#nohup ./align_paired.sh &

# 5. FIltering

## 5.1 Sam to bam

The sam files were converted to bam files: 

%%bash
cd /hydra/chip_RNAi
less sam2bam.sh

#%%bash
#cd ~/../../../../data/ngs/smed/facs_rnaseq_datasets/sexual/chip_RNAi/
#sudo nohup ./sam2bam.sh &

## 5.2 Unique reads and sorting

The sam files were quality-filtered:

%%bash
cd /hydra/chip_RNAi
less samq.sh

#%%bash
#cd ~/../../../../data/ngs/smed/facs_rnaseq_datasets/sexual/chip_RNAi/
#nohup ./samq.sh &

## 5.3 Removing duplicates

Duplicates were removed: 

%%bash
cd /hydra/chip_RNAi/
less rmdup.sh

#%%bash
#cd /hydra/chip_RNAi/
#nohup ./rmdup.sh &

## 5.4 Read retainment 

### 5.4.1 SRR4089759 (h3k4me1_lpt02)

%%bash
cd /hydra/chip_RNAi/
echo $(zcat SRR4089759_1.trimmed.fastq.gz|wc -l)/4|bc

%%bash
cd /hydra/chip_RNAi/
samtools flagstat SRR4089759.sorted.rmdup.bam

print(24472143/(2*30652700))

### 5.4.2 SRR4089760 (h3k4me1_lpt03)

%%bash
cd /hydra/chip_RNAi/
echo $(zcat SRR4089760_1.trimmed.fastq.gz|wc -l)/4|bc

%%bash
cd /hydra/chip_RNAi/
samtools flagstat SRR4089760.sorted.rmdup.bam

print(54329767/(2*76507087))

### 5.4.3 SRR4089771 (input_lpt01)

%%bash
cd /hydra/chip_RNAi/
echo $(zcat SRR4089771_1.trimmed.fastq.gz|wc -l)/4|bc

%%bash
cd /hydra/chip_RNAi/
samtools flagstat SRR4089771.sorted.rmdup.bam

print(17587930/(2*23612971))

### 5.4.4 SRR4089772 (input_lpt02)

%%bash
cd /hydra/chip_RNAi/
echo $(zcat SRR4089772_1.trimmed.fastq.gz|wc -l)/4|bc

%%bash
cd /hydra/chip_RNAi/
samtools flagstat SRR4089772.sorted.rmdup.bam

print(18968785/(2*24840134))

### 3.2.5 SRR4089773 (input_lpt03)

%%bash
cd /hydra/chip_RNAi/
echo $(zcat SRR4089773_1.trimmed.fastq.gz|wc -l)/4|bc

%%bash
cd /hydra/chip_RNAi/
samtools flagstat SRR4089773.sorted.rmdup.bam

print(33320626/(2*46661866))

### 3.2.6 SRR4089846 (h3k4me1_lpt01)

%%bash
cd /hydra/chip_RNAi/
echo $(zcat SRR4089846_1.trimmed.fastq.gz|wc -l)/4|bc

%%bash
cd /hydra/chip_RNAi/
samtools flagstat SRR4089846.sorted.rmdup.bam

print(10294261/(2*13904683))

# 6. Coverage tracks

## 6.1 ChIP-seq H3K4me1 track

Index files were generated for bam files:

%%bash
cd /hydra/chip_RNAi/
less bamindex.sh

%%bash
cd /hydra/chip_RNAi/
ls -l *bai

%%bash
cd /hydra/chip_RNAi/
less bamCov.sh

The mean of RNAi knockdown samples was calculated:

#%%bash
#cd /hydra/chip_RNAi/
#wiggletools mean h3k4me1_lpt01.bw h3k4me1_lpt02.bw h3k4me3_lpt01.bw > h3k4me1_lpt_mean.wig
#wigToBigWig h3k4me1_lpt_mean.wig /hydra/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome h3k4me1_lpt_mean.bw

## 6.2 ChIP-seq H3K4me1 WT vs. lpt(RNAi) track

%%bash
cd /hydra/chip_RNAi/
less h3k4me1RNAilog2.sh

%%bash
cd /hydra/chip_RNAi/
ls h3k4me1RNAilog2.bw

# FINNISHED
