==> notebooks/5-2-MBD34RNAi-H3K4me3ChIP.ipynb <==
# 5-2-MBD34RNAi-H3K4me3ChIP
Jakke Neiro$^1$
1. Aboobaker laboratory, Department of Zoology, University of Oxford

## Contents of notebook

* 1. Introduction
* 2. Data quality
    * 2.1 Data download
    * 2.2 FastQC
        * 2.2.1 SRR4089722 (h3k4me3_gfp01)
        * 2.2.2 SRR4089758 (h3k4me3_gfp02)
        * 2.2.3 SRR4089767 (input_gfp01)
        * 2.2.4 SRR4089768 (input_gfp02)
        * 2.2.5 SRR4089769 (h3k4me3_gfp03)
        * 2.2.6 SRR4089770 (input_gfp03)
        * 2.2.7 SRR4089778 (h3k4me3_lpt01)
        * 2.2.8 SRR4089799 (h3k4me3_lpt02)
        * 2.2.9 SRR4089842 (h3k4me3_lpt03)
* 3. Trimming
    * 3.1 Trimmomatic
* 4. Alignment
* 5. Filtering
    * 5.1 Sam to Bam
    * 5.2 Unique reads and sorting
    * 5.3 Removing duplicates
* 6. Coverage tracks
    * 6.1 ChIP-seq H3K4me3 track
    * 6.2 ChIP-seq H3K4me3 WT vs. lpt(RNAi) track

## Files
* Input: SRR4089842, SRR4089799, SRR4089778, SRR4089770, SRR4089769, SRR4089768, SRR4089767, SRR4089758, SRR4089722
* Output: h3k4me3RNAilog2.bw

# 1. Introduction

The H3K4me3 ChIP-seq data upon MBD3/4 RNAi was inspected, trimmed, and aligned to the planarian genome. Coverage files were generated.

# 2. Data download and FastQC

## 2.1 Data download

The files were donwloaded: 

%%bash
cd /hydra/chip_sexual_organised/k4me3
less sample.list

%%bash
cd /hydra/chip_sexual_organised/k4me3
less data_download_paired.sh

## 2.2 FastQC

FastQC was performed to inspect the quality of the files: 

%%bash
cd /hydra/chip_sexual_organised/k4me3
less fastqc.sh

### 2.2.1 SRR4089722 (h3k4me3_gfp01)

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089722_1_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089722_1_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089722_2_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089722_2_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.2 SRR4089758 (h3k4me3_gfp02)

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089758_1_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd  SRR4089758_1_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089758_2_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd  SRR4089758_2_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.3 SRR4089767 (input_gfp01)

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089767_1_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089767_1_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089767_2_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089767_2_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.4 SRR4089768 (input_gfp02)

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089768_1_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089768_1_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089768_2_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089768_2_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.5 SRR4089769 (h3k4me3_gfp03)

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089769_1_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089769_1_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089769_2_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089768_2_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.6 SRR4089770 (input_gfp03)

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089770_1_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089770_1_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089770_2_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089770_2_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.7 SRR4089778 (h3k4me3_lpt01)

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089778_1_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089778_1_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089778_2_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089778_2_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.8 SRR4089799 (h3k4me3_lpt02)

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089799_1_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089799_1_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089799_2_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089799_2_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.9 SRR4089842 (h3k4me3_lpt03)

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089842_1_fastqc
less summary.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089842_1_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

%%bash
cd /hydra/chip_sexual_organised/k4me3
cd SRR4089842_2_fastqc
less summary.txt

%%bash
cd ~/../../../../data/ngs/smed/facs_rnaseq_datasets/sexual/chip_sexual_organised/k4me3
cd SRR4089842_2_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

# 3. Trimming

## 3.1 Trimmomatic

* SRR4089722_1: TruSeq4 GATCGGAAGAGCACACGTCTGAACTCCAGTCACTGACCA
* SRR4089722_2: GGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGG
* SRR4089758_1: TruSeq12 GATCGGAAGAGCACACGTCTGAACTCCAGTCACCTTGTA
* SRR4089758_2: GGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGG
* SRR4089767_1: TruSeq8 GATCGGAAGAGCACACGTCTGAACTCCAGTCACACTTGA
* SRR4089767_1: GGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGG
* SRR4089768_1: TruSeq4 GATCGGAAGAGCACACGTCTGAACTCCAGTCACTGACCA
* SRR4089768_2: GGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGG
* SRR4089769_1: TruSeq8 GATCGGAAGAGCACACGTCTGAACTCCAGTCACACTTG
* SRR4089769_2: GGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGG
* SRR4089770_1: TruSeq12 GATCGGAAGAGCACACGTCTGAACTCCAGTCACCTTGT
* SRR4089770_2: GGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGG
* SRR4089778_1: TruSeq3 GATCGGAAGAGCACACGTCTGAACTCCAGTCACTTAGGC
* SRR4089778_2: GGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGG
* SRR4089799_1: TruSeq11 GATCGGAAGAGCACACGTCTGAACTCCAGTCACGGCTAC
* SRR4089799_2: GGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGG

%%bash
cd /hydra/chip_sexual_organised/k4me3
less trimmomatic_paired.sh

%%bash
cd /hydra/chip_sexual_organised/k4me3
less fastqc_trimmed.sh

# 4. Alignment

%%bash
cd /hydra/chip_sexual_organised/k4me3
less align_paired.sh

# 5. Filtering

## 5.1 Sam to bam

The sam files were coverted to bam files: 

%%bash
cd /hydra/chip_sexual_organised/k4me3
less sam2bam.sh

## 5.2 Unique reads and sorting

%%bash
cd /hydra/chip_sexual_organised/k4me3
less samq.sh

## 5.3 Removing duplicates

%%bash
cd /hydra/chip_sexual_organised/k4me3
less rmdup.sh

# 6. Coverage

## 6.1 ChIP-seq H3K4me3 track

Index files were generated for bam files: 

%%bash
cd /hydra/chip_sexual_organised/k4me3
less bamindex.sh

#%%bash
#cd /hydra/chip_sexual_organised/k4me3
#nohup ./bamindex.sh &

%%bash
cd /hydra/chip_sexual_organised/k4me3
ls -l *bai

* h3k4me3_gfp01 SRR4089722 and input_gfp01 SRR4089767
* h3k4me3_gfp02 SRR4089758 and input_gfp02 SRR4089768
* h3k4me3_gfp03 SRR4089769 and input_gfp03 SRR4089770
* h3k4me3_lpt01 SRR4089778 and input_lpt01 SRR4089771
* h3k4me3_lpt02 SRR4089799 and input_lpt02 SRR4089772
* h3k4me3_lpt03 SRR4089842 and input_lpt03 SRR4089773

The mean of RNAi knockdown samples was calculated:

%%bash
cd /hydra/chip_sexual_organised/k4me3
less bamCov.sh

#%%bash
#cd /hydra/chip_sexual_organised/k4me3
#wiggletools mean h3k4me3_1.bw h3k4me3_2.bw h3k4me3_3.bw > h3k4me3_mean.wig
#wiggletools mean h3k4me3_lpt1.bw h3k4me3_lpt2.bw h3k4me3_lpt3.bw > h3k4me3_mean_lpt.wig
#wigToBigWig h3k4me3_mean.wig /hydra/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome h3k4me3_mean.bw
#wigToBigWig h3k4me3_mean_lpt.wig /hydra/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome h3k4me3_mean_lpt.bw

## 6.2 ChIP-seq H3K4me3 WT vs. lpt(RNAi) track

%%bash
cd /hydra/chip_sexual_organised/k4me3
less h3k4me3RNAilog2.sh

%%bash
cd /hydra/chip_sexual_organised/k4me3
ls h3k4me3RNAilog2.bw

# FINNISHED
