==> notebooks/5-4-MBD34RNAi-RNA.ipynb <==
# 5-3-MBD34-RNAi-RNA
Jakke Neiro$^1$

1. Aboobaker laboratory, Department of Zoology, University of Oxford

## Contents of notebook
* 1. Introduction
* 2. Data quality
    * 2.1 Data download
    * 2.2 FastQC
        * 2.2.1 SRR4068859 gfp1
        * 2.2.2 SRR4068860 gfp2
        * 2.2.3 SRR4068861 lpt1 RNAi
        * 2.2.4 SRR4068862 lpt2 RNAi
* 3. Trimming
* 4. Alignment: Kallisto
* 5. Target gene table

## Files
* Input: SRR4068859, SRR4068860, SRR4068861, SRR4068862
* Output: kallisto output files, sleuth_ttg_hconf.csv, sleuth_ttg.csv

# 1. Introduction

The RNA-seq data upon MBD3/4 RNAi was inspected and pseudo-aligned to the planarian genome. The transcripts were assigned to genes in the annotation.  

# 2. Data quality

## 2.1 Data download

The files were downloaded:

%%bash
cd /hydra/chip_RNAi/RNAseq
less data_download.sh

%%bash
cd /hydra/chip_RNAi/RNAseq
ls *fastq*

%%bash
cd /hydra/chip_RNAi/RNAseq
less sample.list

#%%bash
#cd /hydra/chip_RNAi/RNAseq
#nohup ./data_download.sh &

#%%bash
#cd /hydra/chip_RNAi/RNAseq
#nohup ./data_download_paired.sh &

## 2.2 FastQC

FastQC was performed on the RNA-seq samples: 

%%bash
cd /hydra/chip_RNAi/RNAseq
less fastqc.sh

#%%bash
#cd /hydra/chip_RNAi/RNAseq
#nohup ./fastqc.sh &

### 2.2.1 SRR4068859 (gfp1)

%%bash
cd /hydra/chip_RNAi/RNAseq
cd SRR4068859_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/RNAseq
cd SRR4068859_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.2 SRR4068860 (gfp2)

%%bash
cd /hydra/chip_RNAi/RNAseq
cd SRR4068860_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/RNAseq
cd SRR4068860_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.3 SRR4068861 (lpt1 RNAi)

%%bash
cd /hydra/chip_RNAi/RNAseq
cd SRR4068861_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/RNAseq
cd SRR4068861_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

### 2.2.4 SRR4068862 (lpt2 RNAi)

%%bash
cd /hydra/chip_RNAi/RNAseq
cd SRR4068862_fastqc
less summary.txt

%%bash
cd /hydra/chip_RNAi/RNAseq
cd SRR4068862_fastqc
grep -A 4 "Overrepresented" fastqc_data.txt

# 3. Trimming

The files were trimmed: 

%%bash
cd /hydra/software/Trimmomatic-0.39/adapters
ls

%%bash
cd /hydra/chip_RNAi/RNAseq
less trimmomatic.sh

#%%bash
#cd /hydra/chip_RNAi/RNAseq
#nohup ./trimmomatic.sh &

%%bash
cd /hydra/chip_RNAi/RNAseq
echo "#!/bin/bash" > fastp.sh
echo "fastp -i SRR4068859_1.fastq.gz -I SRR4068859_2.fastq.gz -o SRR4068859_1.fastp.fastq.gz -O SRR4068859_2.fastp.fastq.gz" >> fastp.sh
echo "fastp -i SRR4068860_1.fastq.gz -I SRR4068860_2.fastq.gz -o SRR4068860_1.fastp.fastq.gz -O SRR4068860_2.fastp.fastq.gz" >> fastp.sh
echo "fastp -i SRR4068861_1.fastq.gz -I SRR4068861_2.fastq.gz -o SRR4068861_1.fastp.fastq.gz -O SRR4068861_2.fastp.fastq.gz" >> fastp.sh
echo "fastp -i SRR4068862_1.fastq.gz -I SRR4068862_2.fastq.gz -o SRR4068862_1.fastp.fastq.gz -O SRR4068862_2.fastp.fastq.gz" >> fastp.sh
chmod +x fastp.sh
less fastp.sh

# 4. Alignment: Kallisto

#%%bash
#cd /hydra/FACS
#gffread -w smeshconf_transcripts.fa -g ~/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4.fa /home/likewise-open/ZOO/quee4048/sexual_genome_annotation_files/sexual_genome_annotation/smes_v2_hconf_SMESG.gtf

%%bash
cd /hydra/chip_RNAi/RNAseq
less kallisto_index.sh

%%bash
cd /hydra/chip_RNAi/RNAseq
less kallisto_index_hconf.sh

#%%bash
#cd /hydra/chip_RNAi/RNAseq
#nohup ./kallisto_index.sh &

#%%bash
#cd /hydra/chip_RNAi/RNAseq
#nohup ./kallisto_index_hconf.sh &

%%bash
cd /hydra/chip_RNAi/RNAseq
ls -l stringtie_transcripts_kallisto
ls -l smeshconf_transcripts_kallisto

The fastq files were pseudo-aligned to the transcripts of the new annotation using Kallisto:

%%bash
cd /hydra/chip_RNAi/RNAseq
less kallisto_align.sh

The fastq files were pseudo-aligned to the transcripts of the Rink annotation using Kallisto:

%%bash
cd /hydra/chip_RNAi/RNAseq
less kallisto_align_hconf.sh

%%bash
cd /hydra/chip_RNAi/RNAseq
less kallisto_align_pairedOct21.sh

%%bash
cd /hydra/chip_RNAi/RNAseq
ls -l *kallisto

%%bash
cd /hydra/chip_RNAi/RNAseq
ls -l *kallisto_hconf

# 5. Target gene table

The transcripts of the Rink annotation were assigned to genes:

import pandas as pd
gffcmp = pd.read_csv("/hydra/sexual_genome_annotation_files/ncrna_Neiro/gffcmp.stringtie_merged.gtf.tmap", sep="\t")
len(gffcmp)

pd.DataFrame({"target_id": gffcmp.iloc[:, 1], "ens_gene": gffcmp.iloc[:, 0]}).to_csv("../FACS/sleuth_ttg_hconf.csv", index=False)

The transcripts of the new annotation were assigned to genes: 

import pandas as pd
sleuth_gene = pd.read_csv("/hydra/FACS/sleuth_gene.csv")

pd.DataFrame({"target_id": sleuth_gene.iloc[:, 1], "ens_gene": sleuth_gene.iloc[:, 21]}).to_csv("../FACS/sleuth_ttg.csv", index=False)

# 6. HISAT2 Alignment

%%bash
cd /hydra/chip_RNAi/RNAseq
echo "#!/bin/bash" > hisat.sh
echo "/hydra/software/hisat2-2.1.0/hisat2 -p 8 --dta --known-splicesite-infile /hydra/neiro_bam/smes_v2_hconf_hisat2_splice_sites -x /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4_hisat2 -U SRR4068859.trimmed.fastq.gz -S SRR4068859.sam" >> hisat.sh 
echo "/hydra/software/hisat2-2.1.0/hisat2 -p 8 --dta --known-splicesite-infile /hydra/neiro_bam/smes_v2_hconf_hisat2_splice_sites -x /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4_hisat2 -U SRR4068860.trimmed.fastq.gz -S SRR4068860.sam" >> hisat.sh
echo "/hydra/software/hisat2-2.1.0/hisat2 -p 8 --dta --known-splicesite-infile /hydra/neiro_bam/smes_v2_hconf_hisat2_splice_sites -x /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4_hisat2 -U SRR4068861.trimmed.fastq.gz -S SRR4068861.sam" >> hisat.sh
echo "/hydra/software/hisat2-2.1.0/hisat2 -p 8 --dta --known-splicesite-infile /hydra/neiro_bam/smes_v2_hconf_hisat2_splice_sites -x /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4_hisat2 -U SRR4068862.trimmed.fastq.gz -S SRR4068862.sam" >> hisat.sh
chmod +x hisat.sh
less hisat.sh

%%bash
cd /hydra/chip_RNAi/RNAseq
echo "#!/bin/bash" > hisat_paired.sh
echo "/hydra/software/hisat2-2.1.0/hisat2 -p 8 --dta --known-splicesite-infile /hydra/neiro_bam/smes_v2_hconf_hisat2_splice_sites -x /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4_hisat2 -1 SRR4068859.trimmed.fastq.gz -2 SRR4068860.trimmed.fastq.gz -S gfp.sam" >> hisat_paired.sh 
echo "/hydra/software/hisat2-2.1.0/hisat2 -p 8 --dta --known-splicesite-infile /hydra/neiro_bam/smes_v2_hconf_hisat2_splice_sites -x /hydra/sexual_genome_annotation_files/sexual_genome_annotation/final_dd_Smed_g4_hisat2 -1 SRR4068861.trimmed.fastq.gz -2 SRR4068862.trimmed.fastq.gz -S lpt.sam" >> hisat_paired.sh
chmod +x hisat_paired.sh
less hisat_paired.sh

%%bash
cd /hydra/chip_RNAi/RNAseq
echo "#!/bin/bash" > samq.sh
echo "samtools view -b -S SRR4068859.sam > SRR4068859.bam" >> samq.sh
echo "samtools view -b -S SRR4068860.sam > SRR4068860.bam" >> samq.sh
echo "samtools view -b -S SRR4068861.sam > SRR4068861.bam" >> samq.sh
echo "samtools view -b -S SRR4068862.sam > SRR4068862.bam" >> samq.sh
chmod +x samq.sh
less samq.sh

%%bash
cd /hydra/chip_RNAi/RNAseq
echo "#!/bin/bash" > samq_p.sh
echo "samtools view -b -S gfp.sam > gfp.bam" >> samq_p.sh
echo "samtools view -b -S lpt.sam > lpt.bam" >> samq_p.sh
chmod +x samq_p.sh
less samq_p.sh

%%bash
cd /hydra/chip_RNAi/RNAseq
echo "#!/bin/bash" > samq_2.sh
echo "samtools sort -@ 8 -o SRR4068859.sorted.bam SRR4068859.bam" >> samq_2.sh
echo "samtools sort -@ 8 -o SRR4068860.sorted.bam SRR4068860.bam" >> samq_2.sh
echo "samtools sort -@ 8 -o SRR4068861.sorted.bam SRR4068861.bam" >> samq_2.sh
echo "samtools sort -@ 8 -o SRR4068862.sorted.bam SRR4068862.bam" >> samq_2.sh
echo "samtools index -b SRR4068859.sorted.bam" >> samq_2.sh
echo "samtools index -b SRR4068860.sorted.bam" >> samq_2.sh
echo "samtools index -b SRR4068861.sorted.bam" >> samq_2.sh
echo "samtools index -b SRR4068862.sorted.bam" >> samq_2.sh
echo "bamCoverage -b SRR4068859.sorted.bam -o SRR4068859.bw --normalizeUsingRPKM" >> samq_2.sh
echo "bamCoverage -b SRR4068860.sorted.bam -o SRR4068860.bw --normalizeUsingRPKM" >> samq_2.sh
echo "bamCoverage -b SRR4068861.sorted.bam -o SRR4068861.bw --normalizeUsingRPKM" >> samq_2.sh
echo "bamCoverage -b SRR4068862.sorted.bam -o SRR4068862.bw --normalizeUsingRPKM" >> samq_2.sh
chmod +x samq_2.sh
less samq_2.sh

%%bash
cd /hydra/chip_RNAi/RNAseq
echo "#!/bin/bash" > samq_2p.sh
echo "samtools sort -@ 8 -o gfp.sorted.bam gfp.bam" >> samq_2p.sh
echo "samtools sort -@ 8 -o SRR4068860.sorted.bam SRR4068860.bam" >> samq_2p.sh
echo "samtools index -b SRR4068859.sorted.bam" >> samq_2p.sh
echo "samtools index -b SRR4068860.sorted.bam" >> samq_2p.sh
echo "bamCoverage -b SRR4068859.sorted.bam -o SRR4068859.bw --normalizeUsingRPKM" >> samq_2p.sh
echo "bamCoverage -b SRR4068860.sorted.bam -o SRR4068860.bw --normalizeUsingRPKM" >> samq_2p.sh
chmod +x samq_2p.sh
less samq_2p.sh

%%bash
cd /hydra/chip_RNAi/RNAseq
ls *bw

%%bash
cd /hydra/chip_RNAi/RNAseq
wiggletools mean gfp1.bw gfp2.bw > gfp_mean.wig
wigToBigWig gfp_mean.wig /hydra/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome gfp_mean.bw

%%bash
cd /hydra/chip_RNAi/RNAseq
wiggletools mean lpt1.bw lpt2.bw > lpt_mean.wig
wigToBigWig lpt_mean.wig /hydra/sexual_genome_annotation_files/sexual_genome_annotation/sizes.genome lpt_mean.bw

%%bash
cd /hydra/chip_RNAi/RNAseq
bigwigCompare -b1 lpt_mean.bw -b2 gfp_mean.bw -o lptlog2.bw

# FINNISHED
